% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconrnaseqParam-class.R, R/epicParam-class.R,
%   R/musicParam-class.R, R/nnlsParam-class.R
\name{deconrnaseqParam}
\alias{deconrnaseqParam}
\alias{epicParam}
\alias{musicParam}
\alias{nnlsParam}
\title{Function to get nnlsParam}
\usage{
deconrnaseqParam(y, z, s = NULL, use.scale = FALSE, return.info = FALSE)

epicParam(y, z, s = NULL, z.var = NULL, return.info = FALSE)

musicParam(
  y,
  z,
  s = NULL,
  sigma = NULL,
  nu = NULL,
  iter.max = NULL,
  eps = NULL,
  return.info = FALSE
)

nnlsParam(y, z, s = NULL, return.info = FALSE)
}
\description{
Function to get nnlsParam

Function to get nnlsParam

Function to get nnlsParam

Function to get nnlsParam
}
