% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexpr_methods.R
\name{analyze_anova}
\alias{analyze_anova}
\title{analyze_anova}
\usage{
analyze_anova(
  sce,
  pheno.df = NULL,
  ngene.sample = 1000,
  assayv = "counts",
  model = "expr ~ celltype * donor",
  return.var = c("percvar"),
  return.sce = TRUE,
  plot.results = TRUE,
  md.name = "anova_results",
  seed.num = 0,
  zoom.panel = TRUE,
  zoom.ylim = c(0, 20),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment or SummarizedExperiment object.}

\item{pheno.df}{Table containing phenotype information. Should include all
variables in provided model except "expr". If NULL, use sce colData.}

\item{ngene.sample}{Number of genes to sample at random.}

\item{assayv}{Vector of names of assays to analyze.}

\item{model}{Character string of the model to use. By default, provides an
interaction between donor and cell type, and the response/dependent variable
is the individual gene.}

\item{return.var}{Vector of variables to return from ANOVA test results (see 
`?get_anova_df` for valid options). By default, returns percentage of sum of
squared variances.}

\item{return.sce}{Whether to return the SingleCellExperiment object with 
results stored as new object in metadata. If FALSE, returns results list.}

\item{plot.results}{Whether to append jitter plots and scatterplots of the
results.}

\item{md.name}{Name of new sce metadata. Ignored if return.sce is FALSE.}

\item{seed.num}{Random seed to set for reproducibility.}

\item{zoom.panel}{Whether to show zoomed panels for jitter plots.}

\item{zoom.ylim}{The y-axis limmits for zoomed panels on jitter plots.}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Additional arguments passed to `anova_scatter_plots()` (see 
`?anova_scatter_plots` for details).}
}
\value{
Either a SingleCellExperiment object with results as a new metadata
entry, or a list of ANOVA results objects.
}
\description{
Main function to perform ANOVA analyses on a series of assays, such as from
a SingleCellExperiment or SummarizedExperiment object.
}
\examples{
# get example data
sce <- random_sce()
sce[["donor"]] <- c(rep("donor1", 2), rep("donor2", 10))

# simple anova
sce <- analyze_anova(sce, model = "expr ~ celltype", ngene.sample = 5)

# anova with batch term -- fails, can't overwrite results metadata
sce <- analyze_anova(sce, model = "expr ~ celltype + donor", 
                     ngene.sample = 5)

# anova with batch term -- success, writing results to new metadata
new.md.name <- "anova_results_batch_model"
sce <- analyze_anova(sce, model = "expr ~ celltype + donor", 
                     ngene.sample = 5, md.name = new.md.name)

# anova with batch interaction term
new.md.name <- "anova_results_batch_interaction_model"
sce <- analyze_anova(sce, model = "expr ~ celltype * donor", 
                     ngene.sample = 5, md.name = new.md.name)

}
\seealso{
get_anova_df, anova_jitter_plots, anova_scatter_plots
}
