% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexpr_methods.R
\name{analyze_dispersion_est}
\alias{analyze_dispersion_est}
\title{analyze_dispersion_est}
\usage{
analyze_dispersion_est(
  sce,
  genes.markerv,
  type.varname = "k2",
  marker.name.str = "top-markers",
  bg.name.str = "bg",
  new.md.name = "dispersion_coefficient_analysis",
  assay.name = "counts_adj",
  num.genes.bg = 100,
  seed.num = 0,
  make.plots = TRUE,
  method.str = "glmGamPoi",
  return.sce = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{genes.markerv}{A vector of marker genes.}

\item{type.varname}{Name of the cell types variable in sce colData columns.}

\item{marker.name.str}{Character string for marker genes.}

\item{bg.name.str}{Character string for background genes.}

\item{new.md.name}{Name of new metadata object in sce. Ignored if 
`return.sce` argument is FALSE.}

\item{assay.name}{Name of assay in sce to analyze.}

\item{num.genes.bg}{Number of random genes to select for background 
estimates.}

\item{seed.num}{Number to set the seed for computational reproducibility.}

\item{make.plots}{Whether to make boxplots and jitter plots of the 
dispersion coefficients by groups from the specified type variable.}

\item{method.str}{Method to fit negative binomial models (see 
`?mexpr_nbcoef()` for details). Defaults to `glmGamPoi` to use `glm_gp()`.}

\item{return.sce}{Whether to return a SingleCellExperiment.}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Additional arguments passed to `plot_dispersion_coeff()`}
}
\value{
Either a SingleCellExperiment object with new metadata, or a list
containing analysis results.
}
\description{
Perform an analysis of dispersion coefficients by fitting negative binomial
models to genes
}
\details{
Analyze the dispersion coefficients from a SingleCellExperiment 
object.

Returned results format is specified by the `return.sce` argument. If this is
TRUE, the SingleCellExperiment object is returned with a new metadata item 
appended called "dispersion_coefficient_analysis". Otherwise, the results are
returned in a list.
}
\examples{
sce <- random_sce()
sce <- analyze_dispersion_est(sce, rownames(sce)[seq(5)], "celltype",
                              assay.name = "counts", num.genes.bg = 10,
                              new.md.name = "dispersion_results")
metadata(sce)[["dispersion_results"]]$plots

}
\seealso{
plot_dispersion_coeff, get_dispersion_coeff_df
}
