% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexpr_methods.R
\name{anova_jitter_plots}
\alias{anova_jitter_plots}
\title{anova_jitter_plots}
\usage{
anova_jitter_plots(dfa, zoom.panel = TRUE, zoom.ylim = c(0, 20))
}
\arguments{
\item{dfa}{Data.frame containing ANOVA results data, from `get_anova_df()`.}

\item{zoom.panel}{Whether to include a zoomed panel. Made by calling 
`ggforce::facet_zoom()`.}

\item{zoom.ylim}{Vector of minimum and maximum y-axis limits for zoomed 
panel. Ignored if the `zoom.panel` argument is FALSE.}
}
\description{
Get jitter scatter plots, grouped by assay, from an anova results data.frame.
}
\examples{
sce <- random_sce()
sce[["donor"]] <- c(rep("donor1", 2), rep("donor2", 8), rep("donor1", 2))
pheno.df <- data.frame(donor = sce[["donor"]], celltype = sce[["celltype"]])
dfa <- get_anova_df(sce = sce, pheno.df = pheno.df)
lggj <- anova_jitter_plots(dfa)
}
