% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexpr_methods.R
\name{anova_scatter_plots}
\alias{anova_scatter_plots}
\title{anova_scatter_plots}
\usage{
anova_scatter_plots(
  dfa.all,
  a1 = "counts",
  a2 = "counts_ds_combat",
  regex.cnv = "^perc\\\\.var\\\\..*"
)
}
\arguments{
\item{dfa.all}{Data.frame of ANOVA results, returned from get_anova_df().}

\item{a1}{Name of first assay, for x-axis.}

\item{a2}{Name of second assay, for y-axis.}

\item{regex.cnv}{Regex pattern to find column names to plot}
}
\description{
anova_scatter_plots
}
