% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{append_groupvar_adj}
\alias{append_groupvar_adj}
\title{convert_sce}
\usage{
append_groupvar_adj(
  set,
  groupvar_pattern = "donor.*",
  assayname = "counts",
  type = "mgvdenom",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{set}{A SummarizedExperimentTypes object}

\item{groupvar_pattern}{String pattern to identify group variables from 
rowData, if present.}

\item{assayname}{Name of original assays data to adjust, as identifiable from
names(assays(set)).}

\item{type}{Adjustment type (options: "mgvdenom").}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Additional arguments passed for specific adjustment type function.}

\item{append_groupvar_adj}{Perform group-wise adjustments on a SummarizedExperimentTypes object. Appends 
group adjustments to rowData and group-adjusted signals to assays.}

\item{rd.method.str}{Character string to append for new assays and rowData 
matrices.}
}
\value{
set object with adjusted marker signal assay data appended.
}
\description{
Manage conversions to a `SummarizedExperimentTypes` object.
}
\examples{
sce <- random_sce()
sce[["donor"]] <- c(rep("donor1", 2), rep("donor2", 8))
set <- set_from_sce(sce, groupvar = "donor")
set <- append_groupvar_adj(set)
}
