% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_cell-scale-factors.R
\name{cell_scale_factors}
\alias{cell_scale_factors}
\title{cell_scale_factors}
\usage{
cell_scale_factors(sce)
}
\arguments{
\item{method}{}
}
\description{
Calculated cell scale factors from a SingleCellExperiment or 
SummarizedExperiment using the specified method.
}
