% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_classes.R
\docType{class}
\name{cellProportionsPredictions-class}
\alias{cellProportionsPredictions-class}
\title{cellProportionsPredictions-class}
\arguments{
\item{predictionsTable}{Table containing cell type predictions.}

\item{cellTypeVector}{Character vector of cell type labels.}

\item{sampleIdVector}{Character vector of sample id labels.}
}
\value{
New cellProportionsPredictions object.
}
\description{
Class for cell type predictions.
}
\details{
Main constructor for class \linkS4class{cellProportionsPredictions}.
}
\examples{
new("cellProportionsPredictions")
predictionsTable <- matrix(sample(100,50),nrow=10)
colnames(predictionsTable) <- paste0("cell_type",seq(ncol(predictionsTable)))
rownames(predictionsTable) <- paste0("sample", seq(nrow(predictionsTable)))
cellProportionsPredictions(predictionsTable)
}
