% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_classes.R
\name{cellProportionsPredictions}
\alias{cellProportionsPredictions}
\title{Make new cellProportionsPredictions object.}
\usage{
cellProportionsPredictions(predictions.table)
}
\arguments{
\item{predictions.table}{Table of cell type predictions.}
}
\value{
cellProportionsPredictions object.
}
\description{
Make new cellProportionsPredictions object.
}
