% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_methods.R
\name{check_coldata}
\alias{check_coldata}
\title{check_coldata}
\usage{
check_coldata(cd, varv)
}
\arguments{
\item{cd}{Matrix of colData.}

\item{varv}{Vector of variable names to check}
}
\value{
list, contains vector of variable values and vector of unique levels
for variables found in cd.
}
\description{
Check coldata columns.
}
