% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor-data-frame-methods.R
\name{check_donordf}
\alias{check_donordf}
\title{check_donordf}
\usage{
check_donordf(df)
}
\arguments{
\item{df}{Data.frame to check.}
}
\value{
Check outcome as boolean; TRUE if `df` passes, or FALSE otherwise.
}
\description{
Checks whether data.frame is a valid simulated donor signals data.frame by
evaluating the column names.
}
