% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconParam-class.R
\docType{class}
\name{deconParam-class}
\alias{deconParam-class}
\alias{DeconParam-class}
\title{deconParam-class}
\description{
Defines the parent class for deconvolution method parameters.
}
\details{
Defines the parent class for deconvolution method parameters. For this class,
the \link{deconvolution} generic performs several useful housekeeping
operations for a standard deconvolution run, including inspecting and 
summarizing properties of the objects z, y, and s.
}
\examples{
deconparam <- new("deconParam")
deconvolution(deconparam)

}
\seealso{
\link{\code{deconvolution}}
}
