% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconrnaseqParam-class.R
\docType{class}
\name{deconrnaseqParam-class}
\alias{deconrnaseqParam-class}
\title{Run DeconRNASeq}
\description{
Runs the DeconRNASeq::DeconRNASeq() deconvolution algorithm.
}
\examples{
# example
lexample <- .get_decon_example_data()
param <- musicParam(s = lexample[["s"]], y = lexample[["y"]], 
z = lexample[["z"]])

# return only predicted proportions
deconvolution(param)
# type1     type2 
# 0.9819837 0.0180163

# return full results
param@return.info <- T
names(deconvolution(param))
# [1] "predictions" "result.info" "metadata"

}
