% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconrnaseqParam-class.R
\name{deconrnaseqParam}
\alias{deconrnaseqParam}
\title{Make new object of class deconrnaseqParam}
\usage{
deconrnaseqParam(y, z, s = NULL, use.scale = FALSE, return.info = FALSE)
}
\arguments{
\item{y}{Bulk mixed signals matrix of samples, which can be matched to single-cell samples.}

\item{z}{Signature matrix of cell type-specific signals. If not provided, can be computed from a
provided ExpressionSet containing single-cell data.}

\item{s}{Cell size factor transformations of length equal to the K cell types to deconvolve.}

\item{use.scale}{Whether to rescale signature matrix data (loigical).}

\item{return.info}{Whether to return metadata and original method outputs with predicted proportions.}
}
\value{
New object of class \linkS4class{deconrnaseqParam}.
}
\description{
Main constructor for class \linkS4class{deconrnaseqParam}.
}
\details{
Takes standard inputs for the DeconRNASeq algorithm
}
