% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisqueParam-class.R
\name{deconvolution,bisqueParam-method}
\alias{deconvolution,bisqueParam-method}
\title{Deconvolution method for bisqueParam}
\usage{
\S4method{deconvolution}{bisqueParam}(object)
}
\value{
Either a vector of predicted proportions, or a list containing predictions, metadata, 
and original outputs.
}
\description{
Main method to access the Bisque deconvolution method from the main lute deconvolution genetic.
}
\details{
Takes an object of class bisqueParam as input, returning a list.
}
