% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconrnaseqParam-class.R
\name{deconvolution,deconrnaseqParam-method}
\alias{deconvolution,deconrnaseqParam-method}
\title{Deconvolution method for class \linkS4class{deconrnaseqParam}}
\usage{
\S4method{deconvolution}{deconrnaseqParam}(object)
}
\value{
Either a vector of predicted proportions, or a list containing 
predictions, metadata, and original outputs.
}
\description{
Main deconvolution method for the \linkS4class{deconrnaseqParam} to run the 
\code{DeconRNASeq::DeconRNASeq()} implementation of the DeconRNASeq algorithm.
}
\references{
Ting Gong and Joseph D. Szustakowski. DeconRNASeq: Deconvolution of 
Heterogeneous Tissue Samples for mRNA-Seq data. (2022), Bioconductor, 
R package version 1.38.0. DOI: 10.18129/B9.bioc.DeconRNASeq
}
