% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolutionParam-class.R
\name{deconvolution,deconvolutionParam-method}
\alias{deconvolution,deconvolutionParam-method}
\title{Deconvolution generic behavior for object of class \linkS4class{deconvolutionParam}}
\usage{
\S4method{deconvolution}{deconvolutionParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{deconvolutionParam}.}
}
\description{
Deconvolution generic behavior for object of class \linkS4class{deconvolutionParam}
}
\details{
Method for behavior of deconvolution generic when called for object of class 
\linkS4class{deconvolutionParam}
}
