% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicParam-class.R
\name{deconvolution,epicParam-method}
\alias{deconvolution,epicParam-method}
\title{Deconvolution method for class \linkS4class{epicParam}}
\usage{
\S4method{deconvolution}{epicParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{epicParam}.}
}
\value{
Either a vector of predicted proportions, or a list containing 
predictions, metadata, and original outputs.
}
\description{
Main deconvolution method for the \linkS4class{epicParam} to run the 
\code{EPIC::EPIC()} implementation of the EPIC algorithm.
}
\examples{
# example
lexample <- lute:::.get_decon_example_data()
param <- epicParam(s = lexample[["s"]], y = lexample[["y"]], z = lexample[["z"]])

# return only predicted proportions
deconvolution(param)

# return full results
param@return.info <- TRUE
names(deconvolution(param))
# [1] "predictions" "result.info" "metadata"

}
\references{
Racle, J., Gfeller, D. (2020). EPIC: A Tool to Estimate the Proportions of 
Different Cell Types from Bulk Gene Expression Data. In: Boegel, S. (eds) 
Bioinformatics for Cancer Immunotherapy. Methods in Molecular Biology, vol 
2120. Humana, New York, NY. https://doi-org.proxy1.library.jhu.edu/10.1007/978-1-0716-0327-7_17

Julien Racle and David Gfeller. EPIC: Estimate the Proportion of Immune and 
Cancer cells. (2022), GitHub, R package version 1.1.5. URL: https://github.com/GfellerLab/EPIC
}
