% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentbulkParam-class.R
\name{deconvolution,independentbulkParam-method}
\alias{deconvolution,independentbulkParam-method}
\title{deconvolution for independentbulkParam-class}
\usage{
\S4method{deconvolution}{independentbulkParam}(object)
}
\description{
Function to perform standard operations prior to deconvolution (a.k.a. "deconvolution prep") for an
object of class independentbulkParam.
}
\details{
Takes an object of independentbulkParam class as input, and returns a list with the filtered/checked/parsed 
experiment objects.
}
