% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music2Param-class.R
\name{deconvolution,music2Param-method}
\alias{deconvolution,music2Param-method}
\title{Deconvolution method for \linkS4class{music2Param}}
\usage{
\S4method{deconvolution}{music2Param}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{music2Param}.}
}
\value{
Either a vector of predicted proportions, or a list containing 
predictions, metadata, and original outputs.
}
\description{
Main method to access the MuSiC2 deconvolution method from the main lute 
deconvolution genetic.
}
\details{
Takes an object of class \linkS4class{music2Param} as input, 
returning a list or vector of predicted cell type proportions.
}
\references{
Fan, Jiaxin. MuSiC2: MuSiC2: cell type deconvolution for multi-condition bulk 
RNA-seq data. (2023) GitHub, R package version 0.1.0. URL: 
https://github.com/Jiaxin-Fan/MuSiC2

Wang, Xuran and Jiaxin Fan. MuSiC: Multi-subject single cell deconvolution. 
(2022) GitHub, R package version 1.0.0. URL: https://github.com/xuranw/MuSiC.

Wang, X., Park, J., Susztak, K. et al. Bulk tissue cell type deconvolution 
with multi-subject single-cell expression reference. Nat Commun 10, 380 
(2019). https://doi.org/10.1038/s41467-018-08023-x

Jiaxin Fan, Yafei Lyu, Qihuang Zhang, Xuran Wang, Mingyao Li, Rui Xiao, 
MuSiC2: cell-type deconvolution for multi-condition bulk RNA-seq data, 
Briefings in Bioinformatics, Volume 23, Issue 6, November 2022, bbac430, 
https://doi-org.proxy1.library.jhu.edu/10.1093/bib/bbac430
}
