% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicParam-class.R
\name{deconvolution,musicParam-method}
\alias{deconvolution,musicParam-method}
\title{Deconvolution method for class \linkS4class{musicParam}}
\usage{
\S4method{deconvolution}{musicParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{music2Param}.}
}
\value{
Either a vector of predicted proportions, or a list containing 
predictions, metadata, and original outputs.
}
\description{
Main deconvolution method for the \linkS4class{musicParam} to run the 
\code{music.basic()} implementation of the MuSiC algorithm.
}
\examples{
# example
lexample <- lute:::.get_decon_example_data()
param <- musicParam(s = lexample[["s"]], y = lexample[["y"]], z = lexample[["z"]])

# return only predicted proportions
deconvolution(param)

# return full results
param@return.info <- TRUE
names(deconvolution(param))

}
\references{
Wang, Xuran and Jiaxin Fan. MuSiC: Multi-subject single cell deconvolution. 
(2022) GitHub, R package version 1.0.0. URL: https://github.com/xuranw/MuSiC.

Wang, X., Park, J., Susztak, K. et al. Bulk tissue cell type deconvolution 
with multi-subject single-cell expression reference. Nat Commun 10, 380 
(2019). https://doi.org/10.1038/s41467-018-08023-x
}
