% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdcParam-class.R
\name{deconvolution,scdcParam-method}
\alias{deconvolution,scdcParam-method}
\title{Deconvolution method for scdcParam}
\usage{
\S4method{deconvolution}{scdcParam}(object)
}
\value{
Either a vector of predicted proportions, or a list containing predictions, metadata, 
and original outputs.
}
\description{
Main method to access the SCDC deconvolution method from the main lute deconvolution genetic.
}
\details{
Takes an object of class scdcParam as input, returning a list or vector of predicted 
cell type proportions.
}
