% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_experiment.R
\name{deconvolution.experiment.info}
\alias{deconvolution.experiment.info}
\title{deconvolution.experiment.info}
\usage{
deconvolution.experiment.info(
  sce,
  s = NULL,
  assay.name = "counts",
  sample.id.variable = "sample.id",
  celltype.variable = "celltype",
  markers.vector = NULL,
  y = NULL,
  experiment.labels = "",
  verbose = FALSE
)
}
\arguments{
\item{sce}{Object of type \linkS4class{SingleCellExperiment}.}

\item{s}{Named numeric vector of cell factors. Used to make the pseudobulk, 
or used by deconvolution.results.table to transform the Z signature matrix.}

\item{assay.name}{Name of expression matrix in \code{sce} assays.}

\item{sample.id.variable}{Name of experiment group variable in \code{sce} 
coldata. Used to group separate deconvolution experiments in a series.}

\item{celltype.variable}{Name of cell type variable in \code{sce} coldata.}

\item{markers.vector}{Vector of marker IDs, used to filter the \code{sce} and 
\code{y} objects prior to deconvolution.}

\item{y}{Bulk/convoluted signals matrix.}

\item{experiment.labels}{Character string vector of labels for the experiment.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
List of length equal to the number of unique sample ids, containing
metadata and key variables for a new deconvolution experiment series.
}
\description{
Get terms for a new deconvolution experiment, startin from an object of type 
\linkS4class{SingleCellExperiment}.
}
\examples{
sce.example <- random_sce()
sce.example[["sample.id"]] <- "sample1"
experiment.info.list <- deconvolution.experiment.info(sce.example)
experiment.info.list$sample1$experiment.labels # ";pseudobulk;"

}
\seealso{
\code{deconvolution.results.table}, \code{deconvolution.experiment}
}
