% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_experiment.R
\name{deconvolution.experiment.permute.groups}
\alias{deconvolution.experiment.permute.groups}
\title{deconvolution.experiment.permute.groups}
\usage{
deconvolution.experiment.permute.groups(
  sce,
  s,
  sample.id.variable = "sample.id",
  celltype.variable = "celltype",
  summary.method = "mean",
  assay.name = "counts",
  deconvolution.algorithm = "nnls",
  experiment.labels = "permutation",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce}{Object of type \linkS4class{SingleCellExperiment}.}

\item{s}{Named numeric vector of cell factors. Used to make the pseudobulk, 
or used by deconvolution.results.table to transform the Z signature matrix.}

\item{sample.id.variable}{Name of experiment group variable in \code{sce} 
coldata. Used to group separate deconvolution experiments in a series.}

\item{celltype.variable}{Name of cell type variable in \code{sce} coldata.}

\item{summary.method}{Statistical method to make the signature matrix from 
an \code{sce} object.}

\item{assay.name}{Name of expression matrix in \code{sce} assays.}

\item{deconvolution.algorithm}{Valid deconvolution algorithm supported by 
the \code{lute} framework (see \code{?lute} for details).}

\item{experiment.labels}{Character string vector of labels for the experiment.}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Additional arguments passed to \code{deconvolution.experiment()} 
(see \code{?deconvolution.experiment} for details).}
}
\value{
List containing the experiment results table and results plot list.
}
\description{
Run a series of permutation experiments.
}
\details{
Run a series of deconvolution experiments, permuting the Z signature 
matrix and Y pseudobulk signals according to the specified experiment group 
id.
}
\examples{
sce.example <- random_sce(num.genes = 100, num.cells = 100)
s <- c("type1" = 3, "type2" = 10)
sce.example[["sample.id"]] <- c(rep("sample1", 10), rep("sample2", 50), rep("sample1", 40))
experiment <- deconvolution.experiment.permute.groups(sce = sce.example, s = s)
sce.example[["sample.id"]] <- "sample1"

}
\seealso{
\code{deconvolution.experiment}, 
\code{deconvolution.results.plots.permutations}
}
