% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_experiment.R
\name{deconvolution.experiment}
\alias{deconvolution.experiment}
\title{deconvolution.experiment}
\usage{
deconvolution.experiment(
  sce,
  y = NULL,
  s = NULL,
  assay.name = "counts",
  sample.id.variable = "sample.id",
  celltype.variable = "celltype",
  deconvolution.algorithm = "nnls",
  typemarker.algorithm = NULL,
  markers.vector = NULL,
  experiment.labels = "",
  verbose = FALSE
)
}
\arguments{
\item{sce}{Object of type \linkS4class{SingleCellExperiment}.}

\item{y}{Bulk/convoluted signals matrix.}

\item{s}{Named numeric vector of cell factors. Used to make the pseudobulk, 
or used by deconvolution.results.table to transform the Z signature matrix.}

\item{assay.name}{Name of expression matrix in \code{sce} assays.}

\item{sample.id.variable}{Name of experiment group variable in \code{sce} 
coldata. Used to group separate deconvolution experiments in a series.}

\item{celltype.variable}{Name of cell type variable in \code{sce} coldata.}

\item{deconvolution.algorithm}{Valid deconvolution algorithm supported by 
the \code{lute} framework (see \code{?lute} for details).}

\item{typemarker.algorithm}{Valid algorithm for marker selection supported 
by the \code{lute} framework, or NULL to skip this.}

\item{markers.vector}{Vector of marker IDs, used to filter the \code{sce} and 
\code{y} objects prior to deconvolution.}

\item{experiment.labels}{Character string vector of labels for the experiment.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
List of experiment groups (smaple IDs) containing variables and 
metadata for a deconvolution experiment series.
}
\description{
Performs a deconvolution experiment.
}
\details{
Perform a full deconvolution experiment.
}
\examples{
sce.example <- random_sce()
sce.example[["sample.id"]] <- "sample1"
full.experiment <- deconvolution.experiment(sce = sce.example, s = c(1, 10))

}
\seealso{
\code{deconvolution.results.summaries}, 
\code{deconvolution.experiment}, \code{plot.deconvolution.results}
}
