% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_experiment.R
\name{deconvolution.results.plots}
\alias{deconvolution.results.plots}
\title{deconvolution.results.plots}
\usage{
deconvolution.results.plots(results.table, verbose = FALSE)
}
\arguments{
\item{results.table}{Experiment results table returned by 
\code{deconvolution.results.table()} (see 
\code{?deconvolution.results.table} for details).}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
List of ggplot2 plots organized by cell types.
}
\description{
Get ggplot2 plots summarizing the results of a deconvolution experiment.
}
\details{
Makes the following results summary plots:

* proportions.scatterplot : Scatter plot of the (x-axis) true and (y-axis) 
predicted cell type proportions.

* abs.error.barplot : Bar plots of (y-axis) absolute errors organized by 
(x-axis) experiment groups. 

* abs.error.jitterbox : Composite jittered points and boxplot overlay of
the (y-axis) absolute errors across experiment groups, organized by (x-axis) 
cell factor adjustment type.
}
\examples{
sce.example <- random_sce()
sce.example[["sample.id"]] <- "sample1"
experiment.info.list <- deconvolution.experiment.info(sce.example)
results.table <- deconvolution.results.table(experiment.info.list)
results.plots <- deconvolution.results.plots(results.table)

}
\seealso{
\code{deconvolution.results.table}, \code{deconvolution.experiment}
}
