% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_experiment.R
\name{deconvolution.results.table}
\alias{deconvolution.results.table}
\title{deconvolution.results.table}
\usage{
deconvolution.results.table(
  experiment.list,
  celltype.variable = "celltype",
  typemarker.algorithm = NULL,
  deconvolution.algorithm = "nnls",
  s = NULL,
  verbose = FALSE
)
}
\arguments{
\item{experiment.list}{An experiment info list output from \code{experiment_info}}

\item{celltype.variable}{Name of cell type variable in \code{sce} and 
\code{z} coldata.}

\item{typemarker.algorithm}{Valid algorithm for marker selection supported 
by the \code{lute} framework, or NULL to skip this.}

\item{deconvolution.algorithm}{Valid deconvolution algorithm supported by 
the \code{lute} framework (see \code{?lute} for details).}

\item{s}{Cell factors for Z signature matrix adjustment. If NULL, use the term 
s.experiment from experiment list sample data.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
Results table, summarizing results and metadata for a deconvolution 
experiment.
}
\description{
Get the results table for a deconvolution experiment.
}
\details{
Performs a deconvolution experiment, aggregating the results into
a results table. Columns in the results table have controlled definitions
and are recognized by downstream functions. The column definitions are:
}
\examples{
sce.example <- random_sce()
sce.example[["sample.id"]] <- "sample1"
experiment.info.list <- deconvolution.experiment.info(sce.example)
results.table <- deconvolution.results.table(experiment.info.list)

}
\seealso{
\code{deconvolution.experiment.info}, \code{deconvolution.experiment}
}
