% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolutionParam-class.R
\docType{class}
\name{deconvolutionParam-class}
\alias{deconvolutionParam-class}
\alias{DeconvolutionParam-class,}
\alias{DeconParam-class,}
\alias{deconParam-class}
\title{deconvolutionParam-class}
\description{
Defines the parent class for deconvolution method parameters.
}
\details{
Defines the parent class for deconvolution method parameters. For this class,
the \link{\code{deconvolution()}} generic performs several useful housekeeping
operations for a standard deconvolution run, including inspecting and 
summarizing properties of the objects z, y, and s.
}
\examples{
param <- new("deconvolutionParam")
deconvolution(param)

}
\seealso{
\link{\code{deconvolution}}
}
