% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{donor_marker_biasexpt}
\alias{donor_marker_biasexpt}
\title{donor_marker_biasexpt}
\usage{
donor_marker_biasexpt(
  offsetv = c(1, 10),
  P = c(0.25, 0.75),
  donor.adj.method = "var_denom",
  gindexv = c(1, 2),
  ndonor = 10,
  ktotal = 2,
  seed.num = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{offsetv}{Vector of donor offset values. This is used to define the
test groups for donor bias experiments. Equal offsets are applied to positive
and negative marker signal distributions.}

\item{P}{Vector of true proportions.}

\item{donor.adj.method}{Method to adjust for donor bias.}

\item{gindexv}{Vector of type indices for the G markers. See `?random_lgv` 
for details.}

\item{ndonor}{Total number of donors to simulate.}

\item{ktotal}{Total K types to simulate.}

\item{seed.num}{Seed value for random sizes in lsv, in case lsv is NULL.}

\item{...}{Arguments passed to function `donoradj()`.}

\item{sd.offset.pos}{Poisson dist mean for randomization of offsets for
positive marker signals.}

\item{sd.offset.neg}{Poisson dist mean for randomization of offsets for
negative marker signals.}

\item{lpv}{List of length num.sim containing true proportions for each 
simulated type. Automatically generated if not provided.}

\item{run.decon}{Whether to run deconvolution experiments, returning 
predictions. Automatically generates values for lpv, lsv if none passed.}
}
\value{
List of experiment results and experiment objects.
}
\description{
Compare predictions with and without donor bias corrections. Parses various
methods for performing the donor bias adjustment, returning predicted 
type proportions for adjusted and unadjusted signature matrices.
}
