% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{donor_marker_sfactorsim}
\alias{donor_marker_sfactorsim}
\title{donor_marker_sfactorsim}
\usage{
donor_marker_sfactorsim(
  gindexv = c(1, 2),
  ndonor = 2,
  ktotal = 2,
  num.sim = 1,
  sd.offset.pos = 5,
  sd.offset.neg = 5,
  lpv = NULL,
  lsv = NULL,
  run.decon = TRUE,
  seed.num = 0,
  plot.title.append = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{gindexv}{Vector of type indices for the G markers. See `?random_lgv` 
for details.}

\item{ndonor}{Total number of donors to simulate.}

\item{ktotal}{Total K types to simulate.}

\item{num.sim}{Number of simulations for deconvolution experiments.}

\item{sd.offset.pos}{Poisson dist mean for randomization of offsets for
positive marker signals.}

\item{sd.offset.neg}{Poisson dist mean for randomization of offsets for
negative marker signals.}

\item{lpv}{List of length num.sim containing true proportions for each 
simulated type. Automatically generated if not provided.}

\item{lsv}{List of length num.sim containing the cell sizes for each type. 
Automatically generated with `seed.num` argument if not provided.}

\item{run.decon}{Whether to run deconvolution experiments, returning 
predictions. Automatically generates values for lpv, lsv if none passed.}

\item{seed.num}{Seed value for random sizes in lsv, in case lsv is NULL.}

\item{...}{Arguments passed to functions `rand_donor_marker_table()`.}
}
\value{
Results of a donor marker experiment, including randomized marker
signal table, results of PCA on marker table, and results and plots of 
deconvolution predictions.
}
\description{
donor_marker_sfactorsim
}
\examples{
donor_marker_sfactorsim()
}
