% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{donoradj}
\alias{donoradj}
\title{donoradj}
\usage{
donoradj(
  df,
  donorv = NULL,
  method = "combat",
  denom_offset = 0.001,
  bounds_thresh = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{df}{A valid `donor.data.frame` object.}

\item{donorv}{Vector of markrer signals (e.g. for a donor, for some summaries 
across donors, etc.).
column names of the format: "donor"+"[0-9]".}

\item{method}{Method to correct for bias. Supports "var_denom", "sd_denom",
"combat".}

\item{denom_offset}{Denominator offset, usually very small, for methods which 
divide by some quantity.}

\item{bounds_thresh}{Threshold for denominator offset. Absolute denominator
values above this threshold are set to equal this value.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
Vector of same length as donorv, containing the bias-adjusted 
values.
}
\description{
Apply some specified bias adjustment to a vector of marker data.
}
\examples{
donordf <- random_donordf()
donorv <- donordf$donor.mean
donoradj(donorv, donordf, method = "combat")

}
\seealso{
biasexpt,
}
