% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{donoradj_combat}
\alias{donoradj_combat}
\title{donoradj_combat}
\usage{
donoradj_combat(df)
}
\arguments{
\item{df}{Data.frame containing marker signals and pheno data.}
}
\value{
madj, matrix of adjusted marker signals.
}
\description{
Use method sva::ComBat() to adjust for donor bias.
}
\examples{
df <- rand_donor_marker_table()
donoradj_combat(df)
}
