% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor-data-frame-methods.R
\name{donordf_from_df}
\alias{donordf_from_df}
\title{donordf_from_df}
\usage{
donordf_from_df(df)
}
\arguments{
\item{df}{Table of type `data.frame`.}
}
\value{
Table of type `donor.data.frame`.
}
\description{
Make a new table of type `donor.data.frame` from a `data.frame`.
}
\details{
Conversion function for `donor.data.frame` class. If missing columns
aren't present in the provided `data.frame`, add them and set their values to
NA.
}
\examples{
df <- data.frame(donor1 = sample(5), marker = rep("marker1", 5))
donordf_from_df(df)

}
