% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donordf_methods.R
\name{donordf_from_mexpr}
\alias{donordf_from_mexpr}
\title{donordf_from_mexpr}
\usage{
donordf_from_mexpr(mexpr, verbose = FALSE)
}
\arguments{
\item{mexpr}{An expression matrix (rows = markers/genes, columns = 
samples/type data).}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
New table of type `donor.data.frame`.
}
\description{
Makes a valid donordf object from an expression matrix (e.g. rows = 
markers/genes, columns = samples/type data)
}
\examples{
df <- rand_donor_marker_table()
madj <- donoradj_combat(df, return.type = "mexpr")
df.adj <- donordf_from_mexpr(mexpr = madj)

}
