% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_typemarker-selection.R
\name{filter_group_markers}
\alias{filter_group_markers}
\title{filter_group_markers}
\usage{
filter_group_markers(
  group.markers,
  minimum.group.overlap.rate = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{minimum.group.overlap.rate}{Minimum amount of overlap by group. Should be a 
fraction of the total groups in the data (e.g. 0.2 means the marker is 
overlapping in 20% of groups).}

\item{verbose}{Whether to show verbose status messages.}

\item{markers.list}{List of gene markers, organized by groups studied (e.g. 
returned from calling \code{marker_overlaps()} (see \code{?marker_overlaps} 
for details).}
}
\value{
List containing marker filter results, including the filtered 
markers list, the overlap info table, and filter metadata.
}
\description{
Get concordant, overlapping markers from a list of markers by group.
}
\details{
Performs the following marker filter steps:
* Uniqueness: Removes duplicate markers across type lists.
* Concordance: Removes markers which aren't specific to the same type across 
groups.
* Overlap: Removes markers below an overlap rate, defined using the 
\code{min.group.overlap.rate} argument.
}
\examples{
sce.example <- random_sce(num.cells = 100)
sce.example[["sample_id"]] <- c(rep("sample1", 10), rep("sample2", 80), rep("sample1", 10))
group.markers <- markers_by_group(sce.example)
filtered.group.markers <- filter_group_markers(group.markers)
}
\seealso{
\code{markers_by_group}, \code{marker_overlaps}
}
