% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon_workflow_functions.R
\name{filter_na_type}
\alias{filter_na_type}
\title{filter_na_type}
\usage{
filter_na_type(
  sce,
  type.variable = "celltype",
  remove.types = TRUE,
  max.gene.na.freq = 0.25,
  max.cells.na.freq = 0.25,
  assayname = "counts",
  append.metadata = TRUE,
  new.metadata.name = "filter.na.type",
  verbose = FALSE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{type.variable}{Name of the type variable in sce colData.}

\item{remove.types}{Whether to remove cells exceeding NA filter from sce 
prior to returning.}

\item{max.gene.na.freq}{The upper threshold for frequency of NA cells 
(see max.na.freq) in a given type, where genes above this count towards the
type gene NA frequency.}

\item{max.cells.na.freq}{The upper threshold for tolerated NA frequency. Types with 
NA frequency exceeding this are flagged for removal.}

\item{assayname}{Name of assay in assays(sce) to check for NAs.}

\item{append.metadata}{Whether to append metadata summary of preprocessing 
parameters and results.}

\item{new.metadata.name}{Name of new metadata object to append (see 
append.metadata argument).}

\item{verbose}{Whether to include verbose status messages.}
}
\value{
scef, filtered SingleCellExperiment with optional appended 
preprocessing metadata.
}
\description{
Filter cell types (or other variable groupings) on frequency of genes or
features with cells exceeding some NA threshold (see max.cells.na.freq).
}
\examples{
# use to filter cell types
sce <- random_sce(na.include = T, na.fract = 0.4)
scef <- filter_na_type(sce, verbose = T)
# can also be used to filter on grouping, such as by donor
colData(sce)$donor <- c(
   rep("donor1", 3), 
   rep("donor2", 5), 
   rep("donor3", 2)
)
scef <- filter_na_cells(sce, type.variable = "donor", 
   new.metadata.name = "filter.na.donor", verbose = T)
}
