% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon_workflow_functions.R
\name{filter_value_cells}
\alias{filter_value_cells}
\title{filter_value_cells}
\usage{
filter_value_cells(
  sce,
  remove.cells = TRUE,
  max.na.freq = 0.25,
  assayname = "counts",
  append.metadata = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{remove.cells}{Whether to remove cells exceeding NA filter from sce 
prior to returning.}

\item{max.na.freq}{The upper threshold for tolerated NA frequency. Cells with 
NA frequency exceeding this are flagged for removal.}

\item{assayname}{Name of assay in assays(sce) to check for NAs.}

\item{append.metadata}{Whether to append metadata summary of preprocessing 
parameters and results.}

\item{verbose}{Whether to include verbose status messages.}
}
\value{
scef, filtered SingleCellExperiment with optional appended 
preprocessing metadata.
}
\description{
Filter cells on NA frequency.
}
\examples{
sce <- random_sce(na.include = T, na.fract = 0.4)
scef <- filter_value_cells(sce, verbose = T)
}
