% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon_workflow_functions.R
\name{filter_value_type}
\alias{filter_value_type}
\title{filter_value_type}
\usage{
filter_value_type(
  sce,
  filter.term = c("zerocount", "NA"),
  type.variable = "celltype",
  remove.types = TRUE,
  max.gene.value.freq = 0.25,
  max.cells.value.freq = 0.25,
  assayname = "counts",
  append.metadata = TRUE,
  new.metadata.name = NULL,
  verbose = FALSE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{filter.term}{Term for filter. Either "zerocount" or "NA".}

\item{type.variable}{Name of the type variable in sce colData.}

\item{remove.types}{Whether to remove cells exceeding NA filter from sce 
prior to returning.}

\item{assayname}{Name of assay in assays(sce) to check for NAs.}

\item{append.metadata}{Whether to append metadata summary of preprocessing 
parameters and results.}

\item{new.metadata.name}{Name of new metadata object to append (see 
append.metadata argument).}

\item{verbose}{Whether to include verbose status messages.}

\item{max.gene.na.freq}{The upper threshold for frequency of NA cells 
(see max.na.freq) in a given type, where genes above this count towards the
type gene NA frequency.}

\item{max.cells.na.freq}{The upper threshold for tolerated NA frequency. Types with 
NA frequency exceeding this are flagged for removal.}
}
\value{
scef, filtered SingleCellExperiment with optional appended 
preprocessing metadata.
}
\description{
Filter cell types (or other variable groupings) on frequency of genes or
features with cells exceeding some frequency threshold (see max.cells.na.freq).
}
\examples{
set.seed(0)
sce <- random_sce(zero.include = T, zero.fract = 0.3)
scef <- filter_value_type(sce, filter.term = "zerocount", verbose = T,
max.gene.value.freq = 0.15)
metadata(scef)$filter.zerocount.by.type$df.type # view new metadata
}
