% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmarkersParam-class.R
\docType{class}
\name{findmarkersParam-class}
\alias{findmarkersParam-class}
\alias{FindmarkersParam-class,}
\alias{findMarkersParam-class}
\title{findmarkersParam-class}
\arguments{
\item{assayName}{Name of expression matrix in SingleCellExperiment assays 
(e.g. "counts").}

\item{singleCellExperiment}{Object of type SingleCellExperiment (see 
\code{?SingleCellExperiment}).}

\item{cellTypeVariable}{Name of cell type variable in SingleCellExperiment 
coldata.}

\item{testType}{Test type (see \code{?findMarkers} for options).}
}
\value{
New object.
}
\description{
class definition for findmarkersParam, which uses scran::findMarkers()
}
\details{
Main constructor for class \linkS4class{findmarkersParam}.
}
\examples{
exampleList <- getDeconvolutionExampleData()
singleCellExperimentExample <- randomSingleCellExperiment()
newParam <- findmarkersParam(singleCellExperiment=singleCellExperimentExample, 
cellTypeVariable="celltype", markersPerType=5)
markers <- typemarkers(newParam)

}
\seealso{
\linkS4class{typemarkersParam}
}
