% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmarkersParam-class.R
\name{findmarkersParam}
\alias{findmarkersParam}
\title{Make new object of class findmarkersParam}
\usage{
findmarkersParam(
  singleCellExperiment,
  assayName = "counts",
  cellTypeVariable = "cellType",
  testType = "wilcox",
  markersPerType = 20,
  returnInfo = FALSE
)
}
\arguments{
\item{singleCellExperiment}{Object of type SingleCellExperiment (see 
\code{?SingleCellExperiment}).}

\item{assayName}{Name of expression matrix in SingleCellExperiment assays 
(e.g. "counts").}

\item{cellTypeVariable}{Name of cell type variable in SingleCellExperiment 
coldata.}

\item{testType}{Test type (see \code{?findMarkers} for options).}

\item{markersPerType}{Number of top markers to get per cell type.}

\item{returnInfo}{Whether to return metadata and original method outputs 
with predicted proportions.}
}
\value{
Object of class \linkS4class{findmarkersParam}
}
\description{
Main constructor for class \linkS4class{findmarkersParam}.
}
\details{
Main class for mapping arguments to the findMarkers method 
implemented as \code{scran::findMarkers()}.
}
\examples{
exampleList <- get_decon_example_data()
singleCellExperimentExample <- random_singleCellExperiment()
newParam <- findmarkersParam(singleCellExperiment=singleCellExperimentExample, 
cellTypeVariable="celltype", markersPerType=5)
markers <- typemarkers(newParam)

}
\seealso{
\linkS4class{typemarkersParam}
}
