% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexpr_methods.R
\name{get_anova_df}
\alias{get_anova_df}
\title{get_anova_df}
\usage{
get_anova_df(
  sce,
  pheno.df,
  ngene.sample = NULL,
  assayv = "counts",
  model = "expr ~ celltype * donor",
  return.var = c("percvar"),
  seed.num = 0,
  verbose = FALSE
)
}
\arguments{
\item{sce}{SingleCellExperiment or SummarizedExperiment object containing
at least one expression assay data matrix.}

\item{pheno.df}{Data.frame containing cell/sample/type phenotype info for the
model matrix. Variables named in argument `model` should be listed here, 
with the exception of "expr", which is defined from the sce expression data 
at runtime.}

\item{ngene.sample}{Number of genes to sample at random. If NULL, fit every
gene in sce.}

\item{model}{Character string of the model to use. By default, provides an
interaction between donor and cell type, and the response/dependent variable
is the individual gene.}

\item{seed.num}{Random seed to set for reproducibility.}

\item{verbose}{Whether to return verbose status messages.}
}
\description{
Get a data.frame of ANOVA statistics.
}
\examples{
sce <- random_sce()
sce[["donor"]] <- c(rep("donor1", 2), rep("donor2", 8), rep("donor1", 2))
pheno.df <- data.frame(donor = sce[["donor"]], celltype = sce[["celltype"]])
dfa <- get_anova_df(sce = sce, pheno.df = pheno.df)

}
