% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_simulation-utilities.R
\name{get_cell_quantities}
\alias{get_cell_quantities}
\title{get_cell_quantities}
\usage{
get_cell_quantities(proportions = c(glial = 0.3, neuron = 0.7), num.min = 100)
}
\arguments{
\item{proportions}{Vector of target proportions by cell type.}

\item{num.min}{Known quantity of cells for minimum proportion cell type(s).}
}
\value{
Vector of cell counts according to specified proportions.
}
\description{
Get the number of each cell type to use across subsample iterations
}
\details{
Gets cell quantities from a provided proportions on minimum count of
least-abundant type according to proportions.
}
\examples{
get_cell_quantities(c("glial" = 0.3, "neuron" = 0.7), 100)
# glial   neuron 
# 100.0000 233.3333 

get_cell_quantities(c("glial" = 0.3, "neuron" = 0.7), 200)
# glial   neuron 
# 200.0000 466.6667 

get_cell_quantities(c("glial" = 0.3, "neuron" = 5), 200)
# glial   neuron 
# 200.000 3333.333 

get_cell_quantities(c("glial" = 0.3, "neuron" = 5, "other" = 0.1), 200)
# glial neuron  other 
# 600  10000    200 

}
