% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{get_celltypes_from_singleCellExperiment}
\alias{get_celltypes_from_singleCellExperiment}
\title{get_celltypes_from_singleCellExperiment}
\usage{
get_celltypes_from_singleCellExperiment(
  singleCellExperiment,
  cellTypeVariable = "celltype"
)
}
\arguments{
\item{singleCellExperiment}{A SingleCellExperiment object.}

\item{cellTypeVariable}{Variable containing cell type labels (e.g. "type1", 
"type2", etc.).}
}
\value{
List of cell type variable metadata and values.
}
\description{
Extract cell type values from SingleCellExperiment.
}
\examples{
lexample <- getDeconvolutionExampleData()

}
