% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{get_decon_example_data}
\alias{get_decon_example_data}
\title{get_decon_example_data}
\usage{
get_decon_example_data(num.bulk.samples = 2, num.markers = 10, num.types = 2)
}
\arguments{
\item{num.bulk.samples}{Number of bulk samples.}

\item{num.markers}{Number of cell type markers.}

\item{num.types}{Number of cell types.}
}
\value{
Example data as list.
}
\description{
Make example data for deconvolution.
}
\examples{
example.data <- get_decon_example_data()

}
