% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{get_decon_example_data_bisque}
\alias{get_decon_example_data_bisque}
\title{get_decon_example_data_bisque}
\usage{
get_decon_example_data_bisque(
  num.bulk.samples = 100,
  num.markers = 1000,
  num.cells = 1000,
  num.types = 2
)
}
\arguments{
\item{num.bulk.samples}{Number of bulk samples.}

\item{num.markers}{Number of cell type markers.}

\item{num.cells}{Number of cells.}

\item{num.types}{Number of cell types.}
}
\value{
Example data as list.
}
\description{
Get example data for Bisque algorithm.
}
\examples{
example.data <- get_decon_example_data()

}
