% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{get_decon_example_data_scdc}
\alias{get_decon_example_data_scdc}
\title{get_decon_example_data_scdc}
\usage{
get_decon_example_data_scdc()
}
\value{
Example data as list.
}
\description{
Get example data for SCDC
}
\examples{
example.data <- get_decon_example_data()

}
