% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_mappings.R
\name{get_deconvolution_predictions}
\alias{get_deconvolution_predictions}
\title{get_deconvolution_predictions}
\usage{
get_deconvolution_predictions()
}
\description{
get_deconvolution_predictions
}
