% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-matrix-methods.R
\name{get_dispersion_coef_df}
\alias{get_dispersion_coef_df}
\title{get_dispersion_coef_df}
\usage{
get_dispersion_coef_df(
  mexpr,
  typev,
  bg.name.str = "bg",
  marker.name.str = "top-markers",
  num.genes.bg = 1000,
  method.str = "glmGamPoi",
  seed.num = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mexpr}{An expression assay matrix (rows = genes/markers, cols = 
cells/samples/types).}

\item{typev}{Vector of length equal to number of columns in mexpr, containing
the cell type labels.}

\item{bg.name.str}{Character string label for background genes.}

\item{marker.name.str}{Character string label for marker genes.}

\item{num.genes.bg}{Number of background genes to select at random.}

\item{method.str}{Method to get dispersion coefficient estimates. This is 
passed to `mexpr_nbcoef` (see `?mexpr_nbcoef` for details).}

\item{seed.num}{Number for random seed.}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Additional arguments passed to `mexpr_nbcoef()`.}
}
\value{
Tall data.frame containing dispersion estimates grouped by types.
}
\description{
get_dispersion_coef_df
}
\seealso{
plot_dispersion_coef, analyze_dispersion_est
}
