% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_methods.R
\name{get_dispersion_data}
\alias{get_dispersion_data}
\title{get_dispersion_data}
\usage{
get_dispersion_data(
  dfstat = NULL,
  mexpr = NULL,
  sce = NULL,
  group.variable = NULL,
  group.vector = NULL,
  assayname = "counts",
  verbose = FALSE
)
}
\arguments{
\item{dfstat}{Table containing statistics for dispersion analysis. Should 
include columns for "mean" and "var".}

\item{mexpr}{Matrix of expression assay data (rows = genes/markers, 
columns = cells/samples).}

\item{sce}{SingleCellExperiment object containing a assays expression matrix
specified by the `assayname` argument.}

\item{group.variable}{Name of sce colData variable containing group labels.}

\item{group.vector}{Vector of group labels of length equal to number of 
columns in mexpr or sce. This is only used if `mexpr` was provided.}

\item{assayname}{Name of expression assay data contained in provided sce
object and/or used in plot title.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
Table dfstat containing statistics (e.g. "mean" and "var" columns)
to use for dispersion analyses.
}
\description{
Get data for dispersion analyses and plots.
}
\examples{
sce.exe <- random_sce() # get a random sce
dfstat <- get_dispersion_data(sce = sce.exe, group.variable = "celltype")

}
\seealso{
plot_dispersion, sce_dispersion
}
