% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_typemarker-selection.R
\name{get_filter_group_markers}
\alias{get_filter_group_markers}
\title{get_filter_group_markers}
\usage{
get_filter_group_markers(
  group.markers,
  minimum.group.overlap.rate = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{minimum.group.overlap.rate}{Minimum amount of overlap by group. Should be a 
fraction of the total groups in the data (e.g. 0.2 means the marker is 
overlapping in 20% of groups).}

\item{verbose}{Whether to show verbose status messages.}

\item{markers.list}{List of gene markers, organized by groups studied (e.g. 
returned from calling \code{marker_overlaps()} (see \code{?marker_overlaps} 
for details).}
}
\value{
List containing markers for removal/filtering, with metadata about the
different filter parameters and type marker selection results.
}
\description{
Get concordant, overlapping markers from a list of markers by group.
}
\details{
Performs the following marker filter steps:
* Uniqueness: Removes duplicate markers across type lists.
* Concordance: Removes markers which aren't specific to the same type across 
groups.
* Overlap: Removes markers below an overlap rate, defined using the 
\code{min.group.overlap.rate} argument.
}
\examples{
set.seed(0)
sce.example <- random_sce(num.cells = 100, num.genes = 1000)
sce.example[["sample_id"]] <- c(rep("sample1", 50), rep("sample2", 50))
group.markers <- markers_by_group(sce.example, markers.per.type = 5)
markers.filter.vector <- get_filter_group_markers(group.markers)
length(markers.filter.vector)

}
\seealso{
\code{markers_by_group}, \code{marker_overlaps}
}
