% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_methods.R
\name{get_groupstat_df}
\alias{get_groupstat_df}
\title{get_groupstat_df}
\usage{
get_groupstat_df(
  exprf,
  groupstat = c("count", "var"),
  summarytype = "colData",
  round.digits = 3
)
}
\arguments{
\item{exprf}{Filtered expression matrix.}

\item{groupstat}{Vector of valid group statistics.}

\item{summarytype}{Type of summary. Either colData or rowData.}

\item{round.digits}{Digits to round.}
}
\value{
Table of type data.frame containing summary statistics.
}
\description{
Get grouped statistics from an expression matrix.
}
\examples{
sce = random_sce(zero.include = TRUE, zero.fract = 0.5)
colData(sce)$donor <- c(rep("donor1", 7), rep("donor2", 3))
exprf <- assays(sce)$counts
get_groupstat_df(exprf)

}
