% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zref_utils.R
\name{get_markers}
\alias{get_markers}
\title{get_markerse}
\usage{
get_markers(marker.method = c("meanratio2"))
}
\arguments{
\item{marker.method}{Supported marker discovery methods.}
}
\description{
Wrapper function for marker discovery procedures.
}
