% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{get_sce_donor_bias}
\alias{get_sce_donor_bias}
\title{get_sce_donor_bias}
\usage{
get_sce_donor_bias(
  donor.bias.coeff = 1,
  mean.vector = NULL,
  num.cells.iter = 10,
  num.genes.iter = 10,
  num.types = 1,
  seed.num = 0
)
}
\arguments{
\item{donor.bias.coeff}{Coefficient of donor bias. Modifies dispersion of 
local negative binomial distributions for sampling.}

\item{mean.vector}{Vector of gene mean values for the sample distribution.
If NULL, uses `seq(1, 30, 1)`.}

\item{num.cells.iter}{Number of cells to simulate for each value in the 
mean.vector.}

\item{num.genes.iter}{Number of genes to simulate for each value in the
mean.vector.}

\item{num.types}{Number of cell types to simulate.}

\item{seed.num}{Token for the random seed.}
}
\value{
New simulated SingleCellExperiment object.
}
\description{
Get a SingleCellExperiment with donor bias incorporated.
}
\details{
The arguments num.cells.iter and num.genes.iter specify the cells 
and genes to simulate for each mean specified by the mean.vector argument. 
Thus for a mean.vector argument of length 10, specifying num.cells.iter = 10 
and num.genes.iter = 5 produces (`10 * 10 = `) 100 cells and (`5 * 10 = `) 50 
genes.

Donor bias is modeled by modifying the dispersion of a negative 
binomial distribution using:

C * M/M^2

Where C is the coefficient of donor bias, and M is the distribution mean. This
formula was chosen because it produces bias that resembles real empirical
donor biases observed for snRNAseq data.
}
