% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{get_set_heatmap}
\alias{get_set_heatmap}
\title{get_set_heatmap}
\usage{
get_set_heatmap(
  set,
  assayname = "logcounts_bytype",
  typevar = NULL,
  groupvar = NULL,
  markertypevar = NULL,
  randcol.seednum = 0,
  scale.color = TRUE,
  verbose = FALSE
)
}
\description{
get_set_heatmap
}
