% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperimentType-methods.R
\name{get_set_pca}
\alias{get_set_pca}
\title{get_set_pca}
\usage{
get_set_pca(
  set,
  assayname = "logcounts_bytype",
  type.variable = NULL,
  group.variable = NULL
)
}
\arguments{
\item{set}{A SummarizedExperimentTypes object or similar.}

\item{assayname}{Name of assays matrix in set object to plot.}

\item{type.variable}{Variable name for type labels in set colData.}

\item{group.variable}{Variable name for type labels in set colData.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
Returns PCA results with plots.
}
\description{
get_set_pca
}
\examples{
sce <- random_sce()
sce[["donor"]] <- c(rep("donor1", 2), rep("donor2", 8))
sce[["typevar"]] <- paste0(sce[["celltype"]], ";", sce[["donor"]])
set <- set_from_sce(sce, group.variable = "donor", type.variable = "typevar")
get_set_pca(set, assayname = "summarized_counts", type.variable = "type")
}
