% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{get_set_plots}
\alias{get_set_plots}
\title{get_set_plots}
\usage{
get_set_plots(
  set,
  assayname = "summarized_counts",
  type.variable = NULL,
  group.variable = NULL,
  mtype.variable = NULL,
  randcol.seednum = 0,
  scale.hmdata = TRUE,
  lplots = c("hm"),
  verbose,
  ...
)
}
\arguments{
\item{lplots}{List of plot types to make. Accepts "pca" and "hm".}
}
\description{
Makes standard plots for SummarizedExperimentTypes and similar objects.
}
