% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zref_utils.R
\name{get_types}
\alias{get_types}
\title{get_types}
\usage{
get_types(
  sce,
  top.markers = 20,
  bind.marker.results = TRUE,
  marker.method = "meanratio2",
  type.method = "mean",
  return.type = "SummarizedExperimentTypes",
  verbose = T,
  ...
)
}
\arguments{
\item{...}{Additional arguments to pass for marker discovery method.}
}
\value{
Either a SummarizedExperimentTypes or or MultiAssayExperiment object.
}
\description{
Get type data from a SingleCellExperiment, returning either an object of 
either class SummarizedExperimentTypes or MultiAssayExperiment.
}
