% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{get_z_from_sce}
\alias{get_z_from_sce}
\title{get_z_from_sce}
\usage{
get_z_from_sce(sce, assay.name = "counts", celltype.variable = "celltype")
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{assay.name}{Name of expression assay type (e.g. "counts").}

\item{celltype.variable}{Name of variable containing cell type labels (e.g. 
"type1", "type2", etc.).}
}
\value{
Matrix of cell summary values (Z reference atlas).
}
\description{
Makes the Z cell atlas reference from a SingleCellExperiment.
}
\examples{
lexample <- get_decon_example_data()
}
