% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{getDeconvolutionExampleData}
\alias{getDeconvolutionExampleData}
\title{getDeconvolutionExampleData}
\usage{
getDeconvolutionExampleData(
  cellScaleFactors = c(1, 10),
  numberBulkSamples = 2,
  numberMarkers = 10,
  numberTypes = 2
)
}
\arguments{
\item{cellScaleFactors}{Vector of cell scale factors}

\item{numberBulkSamples}{Number of bulk samples.}

\item{numberMarkers}{Number of cell type markers.}

\item{numberTypes}{Number of cell types.}
}
\value{
Example data as list.
}
\description{
Make example data for deconvolution.
}
\examples{
exampleData <- getDeconvolutionExampleData()

}
