% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{getDeconvolutionExampleData_scdc}
\alias{getDeconvolutionExampleData_scdc}
\title{getDeconvolutionExampleData_scdc}
\usage{
getDeconvolutionExampleData_scdc()
}
\value{
Example data as list.
}
\description{
Get example data for SCDC
}
\examples{
example.data <- getDeconvolutionExampleData()

}
