% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{getDeconvolutionExampleDataBisque}
\alias{getDeconvolutionExampleDataBisque}
\title{getDeconvolutionExampleData_bisque}
\usage{
getDeconvolutionExampleDataBisque(
  numberBulkSamples = 100,
  numberMarkers = 1000,
  numberCells = 1000,
  numberType = 2
)
}
\arguments{
\item{numberBulkSamples}{Number of bulk samples.}

\item{numberMarkers}{Number of cell type markers.}

\item{numberCells}{Number of cells.}

\item{numberType}{Number of cell types.}
}
\value{
Example data as list.
}
\description{
Get example data for Bisque algorithm.
}
\examples{
exampleData <- getDeconvolutionExampleDataBisque()

}
