% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{getDeconvolutionExampleDataSCDC}
\alias{getDeconvolutionExampleDataSCDC}
\title{getDeconvolutionExampleDataSCDC}
\usage{
getDeconvolutionExampleDataSCDC()
}
\value{
Example data as list.
}
\description{
Get example data for SCDC
}
\examples{
exampleData <- getDeconvolutionExampleDataSCDC()

}
