% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{getDeconvolutionExampleDataSnewColDataC}
\alias{getDeconvolutionExampleDataSnewColDataC}
\title{getDeconvolutionExampleDataSnewColDataC}
\usage{
getDeconvolutionExampleDataSnewColDataC()
}
\value{
Example data as list.
}
\description{
Get example data for SnewColDataC
}
\examples{
exampleData <- getDeconvolutionExampleDataSnewColDataC()

}
