% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{ggpt_donorbias}
\alias{ggpt_donorbias}
\title{ggpt_donorbias}
\usage{
ggpt_donorbias(dfp, method.str = "combat")
}
\arguments{
\item{dfp}{Plot data.frame, containing columns titled: "unadj" (unadjusted 
donor signals), "adj" (adjusted donor signals), "marker" (marker labels), and
"type" (type labels).}

\item{method.str}{Character string describing adjustment method used.}
}
\value{
A ggplot scatterplot object.
}
\description{
Make a scatterplot of the unadjusted and adjusted donor signals.
}
