% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{groupadj_mgvdenom_fromrd}
\alias{groupadj_mgvdenom_fromrd}
\title{groupadj_mgvdenom_fromrd}
\usage{
groupadj_mgvdenom_fromrd(
  set,
  gvariable_pattern = "donor.*",
  assayname = "counts",
  rd.method.str = "mgvdenom",
  verbose = FALSE
)
}
\arguments{
\item{set}{A SummarizedExperimentTypes}

\item{gvariable_pattern}{String pattern to identify group-level summaries from
rowData column labels with grepl.}

\item{assayname}{Name of original assays data to adjust, as identifiable from
names(assays(set)).}

\item{rd.method.str}{Character string to append for new assays and rowData 
matrices.}

\item{verbose}{Whether to return verbose status messages.}
}
\value{
set with updated rowData columns and new assays matrix.
}
\description{
Get adjusted assays and rowadata, calculated as the counts over the mean 
group-wise variances ("mgvdenom"), or:
}
\details{
$$counts/mgvdenom$$
}
