% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentbulkParam-class.R
\docType{class}
\name{independentbulkParam-class}
\alias{independentbulkParam-class}
\title{independentbulkParam-class}
\arguments{
\item{yi}{Mixed signals matrix from bulk samples, independent from primary mixed signals matrix y.}
}
\description{
Class and methods for managing methods requiring independent bulk samples.
}
\examples{
lexample <- .get_decon_example_data()

}
