% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentbulkParam-class.R
\name{independentbulkParam}
\alias{independentbulkParam}
\title{Function to get nnlsParam}
\usage{
independentbulkParam(
  y = NULL,
  yi = NULL,
  z = NULL,
  s = NULL,
  return.info = FALSE
)
}
\description{
Function to get nnlsParam
}
