% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_utils.R
\name{kexpr_sce}
\alias{kexpr_sce}
\title{kexpr_sce}
\usage{
kexpr_sce(
  sce,
  return.lgv = TRUE,
  expr.scale = "logcounts",
  marker.varname = "marker",
  type.varname = "cellType",
  marker.summary = "mean",
  verbose = FALSE
)
}
\arguments{
\item{sce}{Either a SummarizedExperiment or SingleCellExperiment object.}

\item{return.lgv}{Whether to return kexpr summaries as list. If False, returns
as a 2d matrix.}

\item{expr.scale}{Name of expression data obtained from assays(sce), e.g.
logcounts, counts, etc.}

\item{marker.varname}{Marker variable name form rowData(sce).}

\item{type.varname}{Type variable name from colData(sce).}

\item{marker.summary}{Summary function to use, e.g. mean, median, etc.}

\item{vebose}{Whether to return verbose status messages.}
}
\value{
Either a Z signature/reference matrix, or lgv list of marker expression
by type
}
\description{
Get expression summaries by types from SingleCellExperiment or 
SummarizedExperiment objects.
}
\examples{
# example
}
\seealso{
decon_analysis,
}
