% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_metadata.R
\name{lute_supported_deconvolution_algorithms}
\alias{lute_supported_deconvolution_algorithms}
\title{lute_supported_deconvolution_algorithms}
\usage{
lute_supported_deconvolution_algorithms()
}
\value{
Table of supported deconvolution algorithms.
}
\description{
View details about supported deconvolution algorithms.
}
