% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperimentType-methods.R
\name{make_new_assaydata}
\alias{make_new_assaydata}
\title{make_new_assaydata}
\usage{
make_new_assaydata(ma, method = "mean", na.rm = TRUE, verbose = FALSE)
}
\arguments{
\item{ma}{Assay data matrix (cols = samples/cells/donors, rows = genes/loci).
Accepts tables of type data.frame, matrix, or DelayedArray.}

\item{method}{Statistical method to perform summary.}

\item{na.rm}{Option for NA removal.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
Assay data matrix.
}
\description{
Efficiently makes the new assay data by summarizing sce assays with context
sensitivity. Currently supports summary functions using either 
DelayedMartrixStats (if a `DelayedArray` object provided) and Matrix (all 
other inputs).
}
