% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_mappings.R
\name{map_epic}
\alias{map_epic}
\title{map_epic}
\usage{
map_epic(
  arguments,
  method = "EPIC",
  library.name = "EPIC",
  method.arguments = c(bulk = "as.data.frame(Y)", reference = "NA")
)
}
\arguments{
\item{arguments}{List of user-provided arguments.}

\item{method}{Character string of the method.}

\item{library.name}{Name of library to call function from.}

\item{method.arguments}{Arguments required for this method.}
}
\value{
List of data and command character string to parse.
}
\description{
Mapping provided arguments to required arguments for EPIC, using 
defaults for any required arguments not provided
}
