% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_mappings.R
\name{map_music}
\alias{map_music}
\title{map_music}
\usage{
map_music(
  arguments,
  method = "music.basic",
  library.name = "MuSiC",
  method.arguments = c(X = "Z", Y = "Y", S = "S", Sigma = "Sigma", nu = "1e-10",
    iter.max = "100", eps = "0")
)
}
\arguments{
\item{arguments}{List of user-provided arguments.}

\item{method}{Character string of the method.}

\item{library.name}{Name of library to call function from.}

\item{method.arguments}{Arguments required for this method.}
}
\value{
List of data and command character string to parse.
}
\description{
Mapping provided arguments to required arguments for MuSiC, using defaults
for any required arguments not provided
}
