% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_mappings.R
\name{map_nnls}
\alias{map_nnls}
\title{map_nnls}
\usage{
map_nnls(arguments, method = "nnls", method.arguments = c(A = "Z", b = "Y"))
}
\arguments{
\item{arguments}{List of user-provided arguments.}

\item{method}{Character string of the method.}

\item{method.arguments}{Arguments required for this method.}
}
\value{
List of data and command character string to parse.
}
\description{
Mapping provided arguments to required arguments for NNLS, using defaults
for any required arguments not provided
}
