% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_typemarker-selection.R
\name{marker_overlaps}
\alias{marker_overlaps}
\title{marker_overlaps}
\usage{
marker_overlaps(group.markers, marker.filter.vector)
}
\arguments{
\item{group.markers}{Group-specific markers list returned from 
\code{markers_by_group} (see \code{?markers_by_group} for details).}

\item{marker.filter.vector}{Vector of markers to compare (exclude all 
additional markers).}
}
\value{
Table containing marker overlap info.
}
\description{
Gets marker overlap info among groups.
}
\examples{
sce.example <- random_sce(num.cells = 100, num.genes = 100)
sce.example[["sample_id"]] <- c(rep("sample1", 10), rep("sample2", 80), rep("sample1", 10))
group.markers <- markers_by_group(sce.example)
group.overlaps <- marker_overlaps(group.markers, c("gene1", "gene20"))
}
