% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchedbulkParam-class.R
\docType{class}
\name{matchedbulkParam-class}
\alias{matchedbulkParam-class}
\title{matchedbulkParam-class}
\description{
Class and methods for managing methods which allow for bulk sample data to be matched with signatures data.
}
\examples{
lexample <- .get_decon_example_data()

}
