% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanratiosParam-class.R
\docType{class}
\name{meanratiosParam-class}
\alias{meanratiosParam-class}
\alias{MeanratiosParam-class,}
\alias{MeanRatiosParam-class}
\title{nnlsParam-class}
\arguments{
\item{assay.name}{Name of expression matrix in sce assays.}

\item{sce}{Object of type SingleCellExperiment.}

\item{celltype.variable}{Name of cell type variable in sce coldata.}
}
\description{
Uses DeconvoBuddies::get_mean_ratio2()
}
\details{
Main constructor for class \linkS4class{meanratiosParam}.
}
\examples{
lexample <- lute:::.get_decon_example_data()
sce.example <- random_sce()
#new.param <- meanratiosParam(sce = sce.example, celltype.variable = "celltype", markers.per.type = 20)
#markers <- typemarkers(new.param)

}
\seealso{
\linkS4class{typemarkersParam}
}
