% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanratiosParam-class.R
\name{meanratiosParam}
\alias{meanratiosParam}
\title{Make new object of class meanratiosParam}
\usage{
meanratiosParam(
  sce,
  assay.name = "counts",
  celltype.variable = "cellType",
  markers.per.type = 20,
  return.info = FALSE
)
}
\arguments{
\item{sce}{Object of type SingleCellExperiment.}

\item{assay.name}{Name of expression matrix in sce assays.}

\item{celltype.variable}{Name of cell type variable in sce coldata.}

\item{markers.per.type}{Number of top markers to get per cell type.}

\item{return.info}{Whether to return metadata and original method outputs 
with predicted proportions.}
}
\value{
Object of class \linkS4class{meanratiosParam}
}
\description{
Main constructor for class \linkS4class{meanratiosParam}.
}
\details{
Main class for mapping arguments to the mean ratios method 
implemented as \code{DeconvoBuddies::get_mean_ratio2()}.
}
\seealso{
\linkS4class{typemarkersParam}
}
