% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_methods.R
\name{mexpr_downsample}
\alias{mexpr_downsample}
\title{mexpr_downsample}
\usage{
mexpr_downsample(
  mexpr,
  ds.method = "scuttle",
  as.matrix = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{mexpr}{Expression assay matrix.}

\item{ds.method}{Name of the downsampling method to use. The default option
"scuttle" uses `scuttle::downsampleMatrix()`.}

\item{as.matrix}{Whether to convert the downsampled expressiond data to
a matrix object.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
Matrix containing the downsampled expression assays matrix.
}
\description{
Perform downsampling on an expression assay matrix (rows = markers/genes,
columns = cells/samples).
}
\details{
This function provides a wrapper to call downsampling functions for 
an expression assay matrix. The input matrix show have rows corresponding to
genes or markers, and columns corresponding to cells or samples, and it 
should be a matrix-type object.

Downsampling is a technique to select subsets of expression data based on the 
minimum batch-wise observed expression. This can help mitigate batch effects 
in transcriptomics datasets. The operation is a function of the probability
density function or proportions specified for each cell/sample/column in
mexpr. By default, this function uses the formula of min(colsums)/colsums to
define this vector of probabilty densities.
}
\seealso{
sce_dispersion, get_dispersion_data, plot_dispersion
}
