% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{mexpr_from_donordf}
\alias{mexpr_from_donordf}
\title{mexpr_from_donordf}
\usage{
mexpr_from_donordf(df)
}
\arguments{
\item{df}{A donordf type data.frame.}
}
\value{
mexpr, a new expression matrix (rows = genes/markers, columns = 
samples/types)
}
\description{
Make and expression matrix from a donordf data.frame. The donordf contains 
one row per marker;type, while the expression matrix contains one row per
marker and one column per donor;type.
}
