% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon_workflow_functions.R
\name{mexpr_na_freq}
\alias{mexpr_na_freq}
\title{mexpr_na_freq}
\usage{
mexpr_na_freq(mexpr, dim.index = 2)
}
\arguments{
\item{mexpr}{2d matrix.}

\item{dim.index}{Integer for apply index to do NA counts. Either}
}
\value{
list of vectors of NA statistics, including NA counts ("na.countv")
and NA frequencies ("na.freqv")
}
\description{
Get NA frequency from a 2d matrix.
}
