% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellExperiment-methods.R
\name{mexpr_nbcoef}
\alias{mexpr_nbcoef}
\title{mexpr_nbcoef}
\usage{
mexpr_nbcoef(mexpr, method.str = "glmGamPoi", verbose = FALSE)
}
\arguments{
\item{mexpr}{An expression assay matrix (rows = genes/markers, columns = 
cells/samples).}

\item{method.str}{Method used to fit the negative binomial model and compute 
distribution coefficients. If set to "gmlGamPoi" (the default), calls
`glmGamPoi::glm_gp()` for each gene/row in mexpr.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
Results object containing the model coefficients for each gene and
group in mexpr.
}
\description{
Computes the negative binomial distribution coeffcients from an expression
assay matrix
}
\details{
Fits a negative binomial model for each gene in an expression assay
matrix, returning the distribution coefficients. These are useful for running
simulations using empirical estimates of the distribution coefficients, and
for comparing gene-wise expression distributions. This function is called by 
`sce_dispersion()` (see `?sce_dispersion` for details).
}
\seealso{
sce_dispersion
}
