% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music2Param-class.R
\docType{class}
\name{music2Param-class}
\alias{music2Param-class}
\alias{music2Param-class,}
\alias{MuSiC2Param-class,}
\alias{Music2Param-class}
\title{music2Param-class}
\description{
Main constructor for class to manage mappings to the deconvolution method 
function \code{MuSiC::music2_prop()}.
}
\details{
Main constructor for class \linkS4class{music2Param}.
}
\examples{
data <- lute:::.get_decon_example_data_music2()

music2Param()

}
\references{
Fan, Jiaxin. MuSiC2: MuSiC2: cell type deconvolution for multi-condition bulk 
RNA-seq data. (2023) GitHub, R package version 0.1.0. URL: 
https://github.com/Jiaxin-Fan/MuSiC2

Wang, Xuran and Jiaxin Fan. MuSiC: Multi-subject single cell deconvolution. 
(2022) GitHub, R package version 1.0.0. URL: https://github.com/xuranw/MuSiC.

Wang, X., Park, J., Susztak, K. et al. Bulk tissue cell type deconvolution 
with multi-subject single-cell expression reference. Nat Commun 10, 380 
(2019). https://doi.org/10.1038/s41467-018-08023-x

Jiaxin Fan, Yafei Lyu, Qihuang Zhang, Xuran Wang, Mingyao Li, Rui Xiao, 
MuSiC2: cell-type deconvolution for multi-condition bulk RNA-seq data, 
Briefings in Bioinformatics, Volume 23, Issue 6, November 2022, bbac430, 
https://doi-org.proxy1.library.jhu.edu/10.1093/bib/bbac430
}
\seealso{
\linkS4class{deconParam}, \linkS4class{referencebasedParam}, 
\linkS4class{independentbulkParam}
}
