% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music2Param-class.R
\name{music2Param}
\alias{music2Param}
\title{Make an object of class music2Param}
\usage{
music2Param(
  y = NULL,
  yi = NULL,
  z = NULL,
  s = NULL,
  y.eset = NULL,
  sc.eset = NULL,
  assay.name = "counts",
  batch.variable = "SubjectName",
  celltype.variable = "cellType",
  condition.variable = "experiment.condition",
  control.label = "control",
  case.label = "case",
  method.type = "MuSiC2",
  return.info = FALSE
)
}
\arguments{
\item{y}{Bulk mixed signals matrix of samples, which can be matched to single-cell samples.}

\item{yi}{Bulk mixed signals matrix of independent samples, which should not overlap samples in y.}

\item{z}{Signature matrix of cell type-specific signals. If not provided, can be computed from a
provided ExpressionSet containing single-cell data.}

\item{s}{Cell size factor transformations of length equal to the K cell types to deconvolve.}

\item{y.eset}{ExpressionSet of bulk mixed expression signals.}

\item{sc.eset}{ExpressionSet of single-cell transcriptomics data.}

\item{assay.name}{Expression data type (e.g. counts, logcounts, tpm, etc.).}

\item{batch.variable}{Name of variable identifying the batches in sc.eset pData/coldata.}

\item{celltype.variable}{Name of cell type labels variable in sc.eset pData/coldata.}

\item{condition.variable}{Name of variable in y.eset and sc.eset containing condition labels.}

\item{control.label}{Label of control condition samples in condition variable.}

\item{case.label}{Label of case condition samples in condition variable.}

\item{method.type}{Name of method source library to call for music2_prop; either "MuSiC" or "MuSiC2".}

\item{return.info}{Whether to return metadata and original method outputs with predicted proportions.}
}
\value{
Object of class \linkS4class{music2Param}.
}
\description{
Main function to make a new object of class \linkS4class{music2Param}, with
defaults for required arguments.
}
\seealso{
\linkS4class{musicParam}
}
