% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicParam-class.R
\docType{class}
\name{musicParam-class}
\alias{musicParam-class}
\alias{MuSiCParam-class}
\title{musicParam-class}
\description{
Applies the MuSiC::music.basic() implementation of the MuSiC deconvolution 
algorithm.
}
\details{
Main constructor for class \linkS4class{musicParam}.
}
\examples{
# example
lexample <- lute:::.get_decon_example_data()
param <- musicParam(s = lexample[["s"]], y = lexample[["y"]], z = lexample[["z"]])

# return only predicted proportions
deconvolution(param)
# type1     type2 
# 0.6770833 0.3229167

# return full results
param@return.info <- TRUE
names(deconvolution(param))
# [1] "predictions" "result.info" "metadata"

}
\references{
Wang, Xuran and Jiaxin Fan. MuSiC: Multi-subject single cell deconvolution. 
(2022) GitHub, R package version 1.0.0. URL: https://github.com/xuranw/MuSiC.

Wang, X., Park, J., Susztak, K. et al. Bulk tissue cell type deconvolution 
with multi-subject single-cell expression reference. Nat Commun 10, 380 
(2019). https://doi.org/10.1038/s41467-018-08023-x
}
\seealso{
\linkS4class{deconParam}
}
