% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_rnf.R
\name{new_workflow_table}
\alias{new_workflow_table}
\title{new_workflow_table}
\usage{
new_workflow_table(
  sce.names = NULL,
  data.directory = "data",
  true.proportions.filename.stem = "true_proportions_",
  celltype.variable = "celltype",
  table.directory = ".",
  table.filename = "workflow_table.csv",
  save = TRUE,
  overwrite = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sce.names}{Names of SingleCellExperiment files to load.}

\item{data.directory}{Directory containing datasets to load.}

\item{true.proportions.filename.stem}{File name stem of true proportions values.}

\item{celltype.variable}{Name of variable containing cell type labels.}

\item{table.directory}{Directory to write new table.}

\item{table.filename}{Filename of new table to write.}

\item{overwrite}{Whether to overwrite old table files.}

\item{verbose}{Whether to show verbose messages.}
}
\value{
New r-nf_deconvolution compatible table of experiment/run metadata.
}
\description{
Makes a new experiment table for r-nf_deconvolution runs.
}
\details{
Makes and returns/saves a r-nf_deconvolution experiment table. 
Checks for existence of provided files.
}
\examples{
new_workflow_table(save = F)

}
