% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnlsParam-class.R
\name{nnlsParam}
\alias{nnlsParam}
\title{Make new object of class nnlsParam}
\usage{
nnlsParam(
  bulkExpression,
  referenceExpression,
  cellScaleFactors,
  returnInfo = FALSE
)
}
\arguments{
\item{bulkExpression}{Bulk mixed signals matrix of samples, which can be 
matched to single-cell samples.}

\item{referenceExpression}{Signature matrix of cell type-specific signals. 
If not provided, can be computed from a provided \linkS4class{ExpressionSet} 
containing single-cell data.}

\item{cellScaleFactors}{Cell size factor transformations of length equal to 
the K cell types to deconvolve.}

\item{returnInfo}{Whether to return metadata and original method outputs 
with predicted proportions.}
}
\value{
Object of class \linkS4class{nnlsParam}
}
\description{
Main constructor for class \linkS4class{nnlsParam}.
}
\details{
Main parameter class for mapping inputs to the non-negative least 
squares (NNLS) deconvolution algorithm, implemented as \code{nnls::nnls()}.
}
\examples{
exampleList <- getDeconvolutionExampleData()
param <- nnlsParam(cellScaleFactors=exampleList[["cellScaleFactors"]], 
bulkExpression=exampleList[["bulkExpression"]],
referenceExpression=exampleList[["referenceExpression"]])

## return only predicted proportions
deconvolution(param)

# return full results
param@returnInfo <- TRUE
names(deconvolution(param))

}
\seealso{
\linkS4class{referencebasedParam}, \linkS4class{deconvolutionParam}
}
