% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{parseDeconvolutionPredictionsResults}
\alias{parseDeconvolutionPredictionsResults}
\title{parseDeconvolutionPredictionsResults}
\usage{
parseDeconvolutionPredictionsResults(listPred, columnLabels, rowLabels)
}
\arguments{
\item{listPred}{List of cell type proportions predictions.}

\item{columnLabels}{Vector of cell type labels 
(e.g. "type1", "type2", etc.).}

\item{rowLabels}{Vector of sample id labels 
(e.g. "sample1", "sample2", etc.).}
}
\value{
Example data as list.
}
\description{
Gets formatted predicted cell type proportions table from deconvolution 
results list.
}
\examples{
exampleData <- getDeconvolutionExampleData()

}
