% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{pca_bydonortype}
\alias{pca_bydonortype}
\title{pca_bydonortype}
\usage{
pca_bydonortype(dt, verbose = FALSE)
}
\description{
Get analysis results and plots for PCA by donor and type, across markers.
}
