% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_methods.R
\name{plot_dispersion}
\alias{plot_dispersion}
\title{plot_dispersion}
\usage{
plot_dispersion(
  dfstat,
  highlight.markers = NULL,
  hl.color = "black",
  hl.alpha = 0.5,
  show.marker.labels = TRUE,
  hl.padding = 0.5,
  axis.scale = "log10",
  nrow.facet = 1,
  title.str = "Dispersion plot",
  ref.linecol = "red",
  xlab.str = "Mean (log10-scaled)",
  ylab.str = "Variance (log10-scaled)",
  show.smooth = TRUE,
  smooth.linecol = "blue",
  show.points = TRUE,
  point.alpha = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{dfstat}{Table containing summary statistics for mean ("mean") and 
variance ("var").}

\item{highlight.markers}{Vector of markers to highlight in plots. Ignored if
NULL (default).}

\item{hl.color}{Color of points or smooth for highlighted markers. See details.}

\item{hl.alpha}{Transparency of highlighted marker points. See details.}

\item{show.marker.labels}{Whether to show marker labels using 
`ggrepel::geom_text_repel()` (see `?ggrepel::geom_text_repel` for details).}

\item{hl.padding}{Box outline padding amount for when marker labels are 
shown with `show.marker.labels` (see `ggrepel::geom_text_repel` for details).}

\item{nrow.facet}{Number of rows for facet plots. This is only evaluated if 
dfstat contains a column called "group" containing the group labels.}

\item{title.str}{Character string of main plot title.}

\item{ref.linecol}{Color to use for the reference line.}

\item{xlab.str}{Character string of the x-axis title.}

\item{ylab.str}{Character string of the y-axais title.}

\item{show.smooth}{Whether to show the smooth lines using `geom_smooth()` 
with defaults.}

\item{smooth.linecol}{Color to use for smoothed model line.}

\item{show.points}{Whether to show the scatter plot points using 
`geom_point()`.}

\item{point.alpha}{Transparency level for scatter plot points.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
A ggplot2 dispersion plot object.
}
\description{
plot_dispersion
}
\details{
Plot the dispersion (means versus variances) for some summary 
statistics as specified in the `dfstat` object. The `dfstat` object should 
include columns for "mean", "var", and optionally for "marker" and "group". 
it should be tall, meaning that each row corresponds to a marker-group 
summary statistic and there will be >= 1 row per marker if there are >= 1 
unique group labels.

The dispersion plot is highly customizable. It is shown on either the 
identity or the log10 scale (see `axis.scale` argument). One can plot either
the model smooth of the data (e.g. when `show.smooth==TRUE`, the default),
or the scatter plot points (e.g. when `show.points==TRUE`, the default), or 
both. If more than one group is detected, groups are plotted in separate 
facets using `facet_wrap()` with the number of facet rows as specified by
`nrow.facet` (the default is 1 row).

Markers can be highlighted in the dispersion plots. By default, when points
are plotted, separate points are plotted for the markers provided with
`highlight.markers` which are also discovered in dfstat. In this case, the 
highlighted points are shown with the color specified by `hl.color` and 
transparency level from `hl.alpha`. These parameters are also inherited to
show the marker labels with `show.marker.labels` is `TRUE`. When 
`show.points` is `FALSE` but `show.smooth` is `TRUE`, this will instead show 
the smooth of the discovered markers, where the model color is as specified 
in `hl.color`.
}
\examples{
# get dispersion data for plotting
sce.exe <- random_sce() # get a random sce
dfstat <- get_dispersion_data(sce = sce.exe, group.variable = "celltype")

# plot dispersion
#
# 1. plot without highlighting markers
ggds1 <- plot_dispersion(dfstat)
markerv <- dfstat$marker[seq(10)] # get markers for highlighting
#
# 2. highlight some markers without labels
markerv <- dfstat$marker[seq(10)]
ggds2 <- plot_dispersion(dfstat, highlight.markers = markerv, 
                        show.marker.labels = FALSE)
#
# 3. highlight some markers with labels
ggds3 <- plot_dispersion(dfstat, highlight.markers = markerv)
#
# 4. show smooths of all genes and markers, without points
ggds4 <- plot_dispersion(dfstat, highlight.markers = markerv, 
                         show.points = FALSE)

}
\seealso{
get_dispersion_data
}
