% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_simulation-utilities.R
\name{plot_ggpt_bias}
\alias{plot_ggpt_bias}
\title{plot_ggpt_bias}
\usage{
plot_ggpt_bias(dfres, facet = TRUE, verbose = FALSE)
}
\arguments{
\item{dfres}{Simulation series results data.frame.}

\item{facet}{Whether to facet scatterplots for each S condition, if present.}

\item{verbose}{Whether to return verbose status messages.}
}
\value{
ggplot scatterplot object
}
\description{
Makes a ggplot scatterplot of the true proportions (x-axis) by the predicted
proportions (y-axis), where point colors and shapes indicate the type.
}
