% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{plot_ggpt_rmse}
\alias{plot_ggpt_rmse}
\title{plot_ggpt_rmse}
\usage{
plot_ggpt_rmse(dfres, verbose = FALSE)
}
\arguments{
\item{dfres}{Simulation series results data.frame.}

\item{verbose}{Whether to return verbose status messages.}
}
\value{
ggplot scatterplot object
}
\description{
Plots RMSE for zs_transform == FALSE (x-axis) by zs_transform == TRUE 
(y-axis). Note, this requires that the S factor condition varied within
simulations.
}
