% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{plot_ggpt_rmsebyp1}
\alias{plot_ggpt_rmsebyp1}
\title{plot_ggpt_rmsebyp1}
\usage{
plot_ggpt_rmsebyp1(dfres, facet = TRUE, verbose = FALSE)
}
\arguments{
\item{dfres}{Simulation series results data.frame.}

\item{facet}{Whether to facet scatterplots for each S condition, if present.}

\item{verbose}{Whether to return verbose status messages.}
}
\value{
ggplot scatterplot object
}
\description{
Makes a ggplot scatterplot of the RMSE (y-axis) by the true proportion of 
type 1 (x-axis).
}
