% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_simulation-utilities.R
\name{plot_ggvp_rmse}
\alias{plot_ggvp_rmse}
\title{plot_ggvp_rmse}
\usage{
plot_ggvp_rmse(dfres, facet = TRUE, verbose = FALSE)
}
\arguments{
\item{dfres}{Simulation series results data.frame.}

\item{verbose}{Whether to return verbose status messages.}
}
\value{
ggplot violin plot object
}
\description{
Makes a violin plot of the RMSE (y-axis) by experiment condition (x-axis).
}
