% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{predtype}
\alias{predtype}
\title{predtype}
\usage{
predtype(Z, Y, strict_method = "nnls", proportions = TRUE, verbose = FALSE)
}
\arguments{
\item{Z}{Signature/reference matrix, where rows correspond to markers and 
columns correspond to types.}

\item{Y}{Convoluted signals matrix, where rows correspond to markers and 
columns correspond to types.}

\item{strict_method}{Name of a supported strict deconvolution method to use.
See `supported_strict_methods()` for details about supported methods.}

\item{proportions}{Whether to return predicted proportions. If False, returns 
unmodified point prediction outputs.}

\item{verbose}{Whether to show verbose status updates.}
}
\value{
return
}
\description{
Get predictions by type. Takes deconvolution objects and returns vector of 
predictions corresponding to identities of signature/reference matrix types 
(e.g. returns object pres, where type Z[,1] is pres[1], etc.).
}
\examples{
# example
}
\seealso{
decon_results, supported_strict_methods
}
