% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor-subsample_utilities.R
\name{prepare_subsample_experiment}
\alias{prepare_subsample_experiment}
\title{prepare_subsample_experiment}
\usage{
prepare_subsample_experiment(
  sce,
  scale.factor,
  iterations = 10,
  groups.per.iteration = 3,
  method.vector = c("nnls", "music"),
  celltype.variable = "k2",
  group.variable = "Sample",
  assay.name = "counts_adj",
  fraction.cells = 0.25,
  seed.num = 0,
  save.fnstem = "",
  which.save = c("sce", "tp", "ypb", "wt", "li"),
  base.path = "data",
  save.names = list(sce.name = "sce.rda", wt.name = "workflow-table.csv", tp.name =
    "true-proportions.rda", ypb.name = "ypb.rda", li.name = "lindex.rda"),
  verbose = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{scale.factor}{Vector of scale factors corresponding to alphebetical order of
unique type labels in celltype.variable.}

\item{iterations}{Total iterations to perform, per valid method specified in 
the methods argument.}

\item{groups.per.iteration}{Number of groups to sample per iteration. Total 
groups should exceed this value, otherwise data is treated as a single group.}

\item{method.vector}{Vector of valid deconvolution methods to test.}

\item{celltype.variable}{Variable containing cell type labels.}

\item{group.variable}{Variable containing group labels for donor/sample/batch.}

\item{assay.name}{Name of assay in sce to use for pseudobulk.}

\item{fraction.cells}{Fraction of cells, per group and type, to subsample in each iteration.}

\item{seed.num}{Random seed for computational reproducibility.}

\item{save.fnstem}{Character string stem to append to new saved filenames.}

\item{which.save}{Which items to save when running this function. Either "sce" for 
SingleCellExperiment, "tp" for true proportions, "ypb" for pseudobulk, or "wt"
for the `r-nf` compatible workflow table. These objects correspond to filepaths
in save.paths (see details).}

\item{base.path}{Path of directory to store data for workflow run.}

\item{save.names}{List of new file names to save.}
}
\value{
List of iterations data, saving new files as side-effect.
}
\description{
prepare_subsample_experiment
}
\details{
Prepares files for a subsampling experiment to evaluate the impact 
of a group-wise bias effect on deconvolution outcomes. Either real or simulated
data can be provided to the sce argument. This is then randomly subsampled 
where remaining arguments specify details for random iterations. 

The fraction of cells per type and group is specified by the fraction.cells 
argument. To ensure the same number of cells are sampled across iterations,
the final total of cells by type to sample is taken from the group-wise 
minima.

Since it is assumed you will use the same bulk data and true proportions to
evaluate iterations, these data are calculated once using the full sce object.
The pseudobulk dataset is calculated as the produce of the mean signature 
gene expression and the scale.factor cell size scale factor object. The true
cell type proportions correspond to this pseudobulk sample.

The final output files are saved at the base.path specified in save.paths list.
These are suitable for use in an r-nf workflow, or using some other means of
parallelization of the iteration runs.
}
\examples{
sce <- random_sce()
sce[["donor"]] <- "donor1"
iterations.list <- prepare_subsample_experiment(sce, 
celltype.variable = "celltype", group.variable = "donor")
names(iterations.list)

}
