% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_randomized-data.R
\name{proportionsVectorsList}
\alias{proportionsVectorsList}
\title{proportionsVectorsList}
\usage{
proportionsVectorsList(totalCellTypesK = 2, firstCellTypeProportions = NULL)
}
\arguments{
\item{totalCellTypesK}{Total number of cell types to simulate.}

\item{firstCellTypeProportions}{Vector of first cell type proportions. If NULL, uses 
seq(1e-3, 1-1e-3, 1e-3).}
}
\value{
lpv, a list of proportions vectors for simulation iterations.
}
\description{
Get complementary proportions for k types. The first type k1 is the vector of
proportions for the first type. The remaining types up to totalCellTypesK are based on
the reverse of k1. Types k > 1 are assumed to have equal proportions 
complementary to k1.
}
\details{
For k1=c(0, 0.5, 1), totalCellTypesK=2 will generate an additional type with 
proportions c(1, 0.5, 0).

For the same k1 above, totalCellTypesK=3, will generate 2 types with the same 
proportions as c(0.5, 0.25, 0).
}
\examples{
proportionsVectorsList(firstCellTypeProportions=c(0, 0.5, 1))

}
