% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{rand_donor_marker_table}
\alias{rand_donor_marker_table}
\title{rand_donor_marker_table}
\usage{
rand_donor_marker_table(
  ndonor,
  gindexv = c(1, 2),
  mean.offset.pos = 10,
  mean.offset.neg = 2,
  seed.num = 0,
  ...
)
}
\arguments{
\item{ndonor}{Number of donors to simulate.}

\item{mean.offset.pos}{Poisson dist mean for randomization of offsets for
positive marker signals.}

\item{mean.offset.neg}{Poisson dist mean for randomization of offsets for
negative marker signals.}

\item{seed.num}{Token to set the random seed.}

\item{...}{Additional parameters passed to `random_lgv()` to get marker 
signals.}
}
\value{
return
}
\description{
Get a flat table of random donor marker signals by types.
}
\examples{

get_donor_marker_flattable(ndonor = 2, gindexv = c(1,2))

get_donor_marker_flattable(ndonor = 10, gindexv = c(1,1,2))

get_donor_marker_flattable(ndonor = 10, gindexv = c(rep(1, 10), rep(2, 20)))

}
\seealso{
random_lgv
}
