% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomized_data.R
\name{random_lgv}
\alias{random_lgv}
\title{random_lgv}
\usage{
random_lgv(
  gindexv,
  ktotal = 2,
  num.iter = 1,
  lambda.pos = 25,
  lambda.neg = 2,
  seed.num = 0
)
}
\arguments{
\item{ktotal}{Total types to simulate.}

\item{num.iter}{Total simulation iterations.}

\item{randomize.type}{Type of randomization function to use. Can be "poisson"
for Poisson distribution sampling.}

\item{...}{Additional arguments passed to randomization function. For 
randomize.type == "poisson", this is the mpoisson function.}
}
\value{
Listed lgv object containing the randomized marker values across 
types.
}
\description{
Get randomized markers using Poisson distribution sampling. For a given K,
we assume "positive" markers have higher values than for non-K types, and 
thus we sample from 2 different Poisson distributions defined by different
lambda values (e.g. arguments lambda.pos, lambda.neg). WE also use argument 
gindexv to define total markers as length(gindexv) and the marker balance as
relative counts of each type index.
}
\details{
For example, if gindex is c(1, 1, 2), we define 3 total markers, 2 positive
markers for type 1 (negative for type 2) and a single positive marker for 
type 2 (negative for type 1).
}
\examples{
set.seed(0)
random_lgv(gindexv = c(rep(1, 10), rep(2, 5)))
}
