% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomized_data.R
\name{random_sce}
\alias{random_sce}
\title{random_sce}
\usage{
random_sce(...)
}
\arguments{
\item{verbose}{Whether to show verbose status messsages.}
}
\description{
Make a random SingleCellExperiment object.
}
