% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_randomized-data.R
\name{randomSingleCellExperiment}
\alias{randomSingleCellExperiment}
\title{randomSingleCellExperiment}
\usage{
randomSingleCellExperiment(
  numberGenes = 20,
  numberCells = 12,
  numTypes = 2,
  fractionTypes = NULL,
  dispersion = NULL,
  expressionMean = 10,
  naInclude = FALSE,
  naFraction = 0.2,
  zeroInclude = FALSE,
  zeroFraction = 0.2,
  verbose = FALSE,
  seedNumber = 0
)
}
\arguments{
\item{numberGenes}{Number of genes to randomize.}

\item{numberCells}{Numnber of cells to randomize.}

\item{fractionTypes}{Vector of fractions by type.}

\item{dispersion}{Disperison of gene expression. If NULL, uses the mean from 
expressionMean}

\item{expressionMean}{Poisson dist mean for random expression data.}

\item{naInclude}{Whether to include random NA values.}

\item{naFraction}{Fraction of NA values to include.}

\item{zeroInclude}{Whether to include random zero-count values.}

\item{zeroFraction}{Fraction of zero-count values to include.}

\item{verbose}{Whether to show verbose status messages.}

\item{seedNumber}{Seed value for randomization of expression data.}

\item{numberTypes}{Number of cell types to annotate.}
}
\value{
New randomized SingleCellExperiment object.
}
\description{
Make a random object of type SingleCellExperiment. Uses the negative binomial 
distribution to randomly generate gene expression data for simulated cells.
}
\examples{
singleCellExperiment <- randomSingleCellExperiment()

}
