% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencebasedParam-class.R
\docType{class}
\name{referencebasedParam-class}
\alias{referencebasedParam-class}
\title{referencebasedParam-class}
\description{
Class and methods for managing reference-based deconvolution methods.
}
\examples{
lexample <- .get_decon_example_data()

}
