% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencebasedParam-class.R
\name{referencebasedParam}
\alias{referencebasedParam}
\title{Make a new object of class \linkS4class{referencebasedClass}}
\usage{
referencebasedParam(y, z, s = NULL, return.info = FALSE)
}
\value{
New object of class \linkS4class{referencebasedClass}
}
\description{
Make a new object of class \linkS4class{referencebasedClass}
}
