% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencefreeParam-class.R
\docType{class}
\name{referencefreeParam-class}
\alias{referencefreeParam-class}
\title{referencefreeParam-class}
\description{
Class and methods for managing reference-free deconvolution methods.
}
\details{
Reference-free deconvolution methods requently lack a signature 
matrix or use alternative means of prediction such as machine learning 
models. This parent class is meant to manage methods of this type.
}
\examples{
lexample <- lute:::.get_decon_example_data()
new("referencefreeParam", y = lexample$y)

}
