% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{referenceFromSingleCellExperiment}
\alias{referenceFromSingleCellExperiment}
\title{referenceFromSingleCellExperiment}
\usage{
referenceFromSingleCellExperiment(
  singleCellExperiment,
  assayName = "counts",
  cellTypeVariable = "celltype"
)
}
\arguments{
\item{singleCellExperiment}{A SingleCellExperiment object.}

\item{assayName}{Name of expression assay type (e.g. "counts").}

\item{cellTypeVariable}{Name of variable containing cell type labels (e.g. 
"type1", "type2", etc.).}
}
\value{
Matrix of cell summary values (Z reference atlas).
}
\description{
Makes the Z cell atlas reference from a SingleCellExperiment.
}
\examples{
lexample <- getDeconvolutionExampleData()
}
