% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R, R/sim_utils_sce.R
\name{results_plots}
\alias{results_plots}
\title{results_plots}
\usage{
results_plots(dfres, verbose = FALSE)

results_plots(dfres, verbose = FALSE)
}
\arguments{
\item{dfres}{Data.frame of deconvolution simulation results.}

\item{verbose}{Whether to show verbose status updates.}
}
\value{
List of ggplot2 objects analyzing deconvolution simulation results, 
including scatter plots and a violin plot.

List of ggplot2 objects analyzing deconvolution simulation results, 
including scatter plots and a violin plot.
}
\description{
Makes standard plots to analyze deconvolution simulation results.

Makes standard plots to analyze deconvolution simulation results.
}
