% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_rmse.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(proportionsTrue, proportionsPred, summaryType = "mean")
}
\arguments{
\item{proportionsTrue}{cell type proportions taken as true}

\item{proportionsPred}{cell type proportions taken as false}

\item{summaryType}{Toggle summary type (either "mean" or "median")}
}
\value{
single numeric
}
\description{
Takes 2 vectors of numerics
}
\details{
Calculates the root mean squared error (RMSE) for specified true and 
predicted cell type proportions.


Function does not distinguish between true and predicted status, variable 
labels provided for convenience.
}
\examples{
proportionsVectorPred <- seq(1e-10,2e-10,1e-11)
proportionsVectorTrue <- rev(proportionsVectorPred)
rmse(proportionsVectorTrue, proportionsVectorPred)

}
