% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_rmse.R
\name{rmseTest}
\alias{rmseTest}
\title{rmseTest}
\usage{
rmseTest(firstVector, secondVector)
}
\arguments{
\item{firstVector}{First numeric vector.}

\item{secondVector}{Second numeric vector.}
}
\value{
Single numeric value
}
\description{
Takes 2 vectors of numerics
}
\details{
Tests the rmse function for rounding imprecision.

Function to test RMSE values (`./unitTests/test_rmse.R`).
}
\examples{
proportionsVectorPred <- seq(1e-10,2e-10,1e-11)
proportionsVectorTrue <- rev(proportionsVectorPred)
rmseTest(proportionsVectorTrue, proportionsVectorPred)
}
