% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdcParam-class.R
\name{scdc_basis_eset}
\alias{scdc_basis_eset}
\title{Make SCDC basis matrix as an ExpressionSet object}
\usage{
scdc_basis_eset(sc.eset, ct.sub, ct.varname, min.sum = 0)
}
\arguments{
\item{sc.eset}{ExpressionSet of single cell data.}

\item{ct.sub}{Cell types vector.}

\item{ct.varname}{Variable name for cell type labels in sc.eset.}

\item{min.sum}{Minimum total/sums expression threshold for marker inclusion.}
}
\value{
ExperimentSet containing SCDC basis matrix.
}
\description{
Make SCDC basis matrix as an ExpressionSet object
}
\details{
This is an intermediate data object used by the SCDC algorithm.
}
