% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdcParam-class.R
\docType{class}
\name{scdcParam-class}
\alias{scdcParam-class}
\alias{SCDCParam-class,}
\alias{ScdcParam-class,}
\alias{ScdcParam-class}
\title{scdcParam-class}
\description{
Main constructor for class to manage mappings to the deconvolution method function \code{SCDC::SCDC_prop()}.
}
\details{
Main constructor for class \linkS4class{scdcParam}.
}
\examples{
lexample <- lute:::.get_decon_example_data()

}
\references{
Meichen Dong. SCDC: Bulk RNA-seq deconvolution by scRNA-seq with multi-reference 
datasets. (2023) GitHub, R package version 0.0.0.9000. URL: https://github.com/meichendong/SCDC

Meichen Dong, Aatish Thennavan, Eugene Urrutia, Yun Li, Charles M Perou, Fei 
Zou, Yuchao Jiang, SCDC: bulk gene expression deconvolution by multiple 
single-cell RNA sequencing references, Briefings in Bioinformatics, Volume 
22, Issue 1, January 2021, Pages 416–427, 
https://doi-org.proxy1.library.jhu.edu/10.1093/bib/bbz166
}
\seealso{
\linkS4class{deconParam}, \linkS4class{referencebasedParam}, \linkS4class{independentbulkParam}
}
