% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-matrix_utilities.R
\name{sce_append_markers}
\alias{sce_append_markers}
\title{sce_append_markers}
\usage{
sce_append_markers(
  sce,
  marker.varname = "typeMarker",
  marker.method = c("meanratio2")
)
}
\arguments{
\item{marker.method}{Supported marker discovery methods.}

\item{verbose}{Whether to show verbose status messages.}
}
\description{
Wrapper function for marker discovery procedures. Accepts and returns a 
SingleCellExperiment, with marker labels appended.
}
