% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_methods.R
\name{sce_dispersion}
\alias{sce_dispersion}
\title{sce_dispersion}
\usage{
sce_dispersion(
  expr.data,
  group.data = NULL,
  plot.dispersion = TRUE,
  highlight.markers = NULL,
  assayname = "counts",
  downsample = TRUE,
  ds.method = "scuttle",
  get.nbstat = FALSE,
  method.nbstat = "glmGamPoi",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{expr.data}{Either a matrix, a SingleCellExperiment, or 
SummarizedExperiment object.}

\item{group.data}{Either a colData variable name (if expr.data is of type 
SingleCellExperiment or SummarizedExperiment), or a vector of labels of 
length equal to the number of columns in expr.data (if expr.data is a 
matrix).}

\item{plot.dispersion}{Whether to make plots of dispersion using 
`plot_dispersion()` (see `?plot_dispersion` for details).}

\item{highlight.markers}{Vector of marker identifiers to highlight in 
dispersion plots. Ignored if NULL (default).}

\item{assayname}{Name of the expression assay matrix. Only used if expr.data
is a SingleCellExperiment or SummarizedExperiment.}

\item{downsample}{Whether to perform downsampling by calling 
`mexpr_downsample()` (see `?mexpr_downsample` for details).}

\item{ds.method}{Name of the downsampling method to use (only used if 
`downsample` is `TRUE`).}

\item{get.nbstat}{Compute the negative binomial distribution coefficients 
for each gene/marker. Calls the function `mexpr_nbcoef()` (see 
`?mexpr_nbcoef` for details).}

\item{method.nbstat}{Method to compute the negative binomial distribution
coefficients. Should be a valid method recognized by the function 
`mexpr_nbcoef()` (see `?mexpr_nbcoef` for details).}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Additional arguments passed to `plot_dispersion()`.}
}
\value{
List of dispersion results, including the summary statistics table
and the dispersion plot.
}
\description{
Perform dispersion analysis for a SingleCellExperiment object.
}
\details{
This is the main function to manage a dispersion analysis for a 
SingleCellExperiment, SummarizedExperiment, or matrix of expression assay
data. This generates and checks the group-level summary statistics, then 
produces plots of gene/marker-wise points or model smooths showing the 
mean and variance relationship. The coefficients for the negative binomial 
distribution can also be generated for each gene in each provided group.
}
\examples{
# get some random example sce data
sce.exe <- random_sce()

# analyze dispersion
#
# get dispersion results with defaults
ld1 <- sce_dispersion(sce.exe, group.data = "celltype", verbose = TRUE)
#
# show marker highlights in plots
ld2 <- sce_dispersion(sce.exe, group.data = "celltype", verbose = TRUE,
highlight.markers = rownames(sce.exe)[seq(10)])
# 
# get negative binomial coefficients
ld3 <- sce_dispersion(sce.exe, get.nbstat = TRUE)

}
\seealso{
get_dispersion_data, plot_dispersion, mexpr_downsample, mexpr_nbcoef
}
