% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_preprocessing.R
\name{sce_preprocess_groups}
\alias{sce_preprocess_groups}
\title{sce_preprocess_groups}
\usage{
sce_preprocess_groups(
  sce,
  group.variable = "sample.id",
  celltype.variable = "celltype",
  assay.name = "counts",
  new.assay.stem = "adjusted",
  downsample = TRUE,
  zscale.threshold = NULL,
  negative.to.zero.expression = TRUE,
  return.type = "list",
  verbose = FALSE
)
}
\arguments{
\item{group.variable}{Variable containing group labels ("sample.id").}

\item{celltype.variable}{Variable containing cell type labels ("celltype").}

\item{assay.name}{Name of assay to adjust ("counts").}

\item{new.assay.stem}{New assay character string stem ("adjusted").}

\item{downsample}{Whether to perform downsampling on group-type cell amounts (TRUE).}

\item{zscale.threshold}{Minimum threshold to filter whole groups on their 
z-scale values for group-type cell amounts (NULL to skip).}

\item{negative.to.zero.expression}{Whether to set negative expression values to 
zero after adjustment (TRUE).}

\item{return.type}{Either "list", including summary statistics metadata, or "sce" for
SingleCellExperiment without metadata.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
SingleCellExperiment object with group-adjusted expression.
}
\description{
Perform group adjustment on a SingleCellExperiment.
}
\details{
Perform group adjustment on a SingleCellExperiment object. Includes options to
perform downsampling and filtering of whole groups on the z-scale of group-type cell 
amounts.
}
\examples{
sce.example <- random_sce(num.genes = 1000, num.cells = 1000)
sce.example[["sample.id"]] <- c(rep("sample1", 800), rep("sample2", 200))
list.preprocess <- sce_preprocess_groups(sce.example)
names(list.preprocess)
names(list.preprocess$metadata)
names(assays(sce))

}
