% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_conversions.R
\name{sce_to_eset}
\alias{sce_to_eset}
\title{sce_to_eset
Convert SingleCellExperiment to ExpressionSet.}
\usage{
sce_to_eset(sce, assay.name = "counts")
}
\arguments{
\item{sce}{Object of type SingleCellExperiment (see 
\code{?SingleCellExperiment}).}

\item{assay.name}{Name of assay to store in new eset.}
}
\value{
ExpressionSet.
}
\description{
sce_to_eset
Convert SingleCellExperiment to ExpressionSet.
}
\examples{
sce <- random_sce()
sce_to_eset(sce, "counts")

}
