% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_conversions.R
\name{sce_to_se}
\alias{sce_to_se}
\title{sce_to_se
Convert SingleCellExperiment to SummarizedExperiment.}
\usage{
sce_to_se(singleCellExperiment)
}
\arguments{
\item{singleCellExperiment}{Object of type SingleCellExperiment (see 
\code{?SingleCellExperiment}).}
}
\value{
SummarizedExperiment.
}
\description{
sce_to_se
Convert SingleCellExperiment to SummarizedExperiment.
}
\examples{
sce <- randomSingleCellExperiment()
sce_to_se(sce)
}
