% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{set_from_sce}
\alias{set_from_sce}
\title{set_from_sce}
\usage{
set_from_sce(
  sce,
  method = "mean",
  type.variable = "celltype",
  assayname = "counts",
  verbose = FALSE
)
}
\arguments{
\item{sce}{A `SingleCellExperiment` object.}

\item{method}{Method to summarize assays on types.}

\item{type.variable}{Variable containing the type-level labels.}

\item{assayname}{Name of assays data in assays(sce).}

\item{verbose}{Whether to show verbose status messages.
groups (see `?sce_groupstat` for details).}
}
\value{
`SummarizedExperimentTypes` object.
}
\description{
Make an object of class `SummarizedExperimentTypes` from an object of class
`SingleCellExperiment` or `SummarizedExperiment`.
}
\examples{
set <- set_from_sce(random_sce()); show(set)
}
\seealso{
sce_groupstat
}
