% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolutionParam-class.R
\name{show,deconvolutionParam-method}
\alias{show,deconvolutionParam-method}
\title{Show generic behavior for object of class \linkS4class{deconvolutionParam}}
\usage{
\S4method{show}{deconvolutionParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{deconvolutionParam} (see 
\code{?deconvolutionParam}).}
}
\value{
Shows object summaries.
}
\description{
Show generic behavior for object of class \linkS4class{deconvolutionParam}
}
\details{
Method for behavior of show generic when called for object of class 
\linkS4class{deconvolutionParam}
}
\examples{
param <- new("deconvolutionParam")
deconvolution(param)

}
