% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmarkersParam-class.R
\name{show,findmarkersParam-method}
\alias{show,findmarkersParam-method}
\title{Show generic behavior for object of class \linkS4class{findmarkersParam}}
\usage{
\S4method{show}{findmarkersParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{findmarkersParam} (see 
\code{?findmarkersParam}).}
}
\value{
Shows object summaries.
}
\description{
Show generic behavior for object of class \linkS4class{findmarkersParam}
}
\details{
Method for behavior of show generic when called for object of class 
\linkS4class{findmarkersParam}
}
\examples{
exampleList <- getDeconvolutionExampleData()
singleCellExperimentExample <- randomSingleCellExperiment()
newParam <- findmarkersParam(singleCellExperiment=singleCellExperimentExample, 
cellTypeVariable="celltype", markersPerType=5)
markers <- typemarkers(newParam)

}
