% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentbulkParam-class.R
\name{show,independentbulkParam-method}
\alias{show,independentbulkParam-method}
\title{Method for \linkS4class{independentbulkParam}}
\usage{
\S4method{show}{independentbulkParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{independentbulkParam} (see 
\code{?independentbulkParam}).}
}
\value{
Shows object summaries.
}
\description{
Method for \linkS4class{independentbulkParam}
}
\details{
Display data summaries for an object of class 
\linkS4class{independentbulkParam}.
}
\examples{
new("independentbulkParam")

}
