% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnlsParam-class.R
\name{show,nnlsParam-method}
\alias{show,nnlsParam-method}
\title{Show generic behavior for object of class \linkS4class{nnlsParam}}
\usage{
\S4method{show}{nnlsParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{nnlsParam} (see 
\code{?nnlsParam}).}
}
\value{
Shows object summaries.
}
\description{
Show generic behavior for object of class \linkS4class{nnlsParam}
}
\details{
Method for behavior of show generic when called for object of class 
\linkS4class{nnlsParam}
}
\examples{
exampleList <- getDeconvolutionExampleData()
param <- nnlsParam(
cellScaleFactors=exampleList[["cellScaleFactors"]], 
bulkExpression=exampleList[["bulkExpression"]],
referenceExpression=exampleList[["referenceExpression"]])

## return only predicted proportions
deconvolution(param)

# return full results
param@returnInfo <- TRUE
names(deconvolution(param))

}
