% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencebasedParam-class.R
\name{show,referencebasedParam-method}
\alias{show,referencebasedParam-method}
\title{Show generic behavior for object of class referencebasedParam}
\usage{
\S4method{show}{referencebasedParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{referencebasedParam}.}
}
\value{
Prints data summary messages to console.
}
\description{
Show generic behavior for object of class referencebasedParam
}
\examples{
lexample <- get_decon_example_data()
referencebasedParam(y = lexample$y, z = lexample$z, s = lexample$s)
}
