% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{signature_matrix_from_sce}
\alias{signature_matrix_from_sce}
\title{signature_matrix_from_sce}
\usage{
signature_matrix_from_sce(
  sce,
  celltype.variable = "celltype",
  summary.method = "mean",
  assay.name = "counts"
)
}
\arguments{
\item{sce}{An object of type \linkS4class{SingleCellExperiment}.}

\item{celltype.variable}{Variable name for cell type labels in \code{sce} 
coldata.}

\item{summary.method}{Summary statistic function to use.}

\item{assay.name}{Name of expression matrix in \code{sce} assays.}
}
\value{
New Z signature matrix.
}
\description{
Calculate a Z signature matrix from object of type 
\linkS4class{SingleCellExperiment}.
}
\details{
Calculate a Z signature matrix from object of type 
\linkS4class{SingleCellExperiment}.
}
\examples{
sce.example <- random_sce()
signature_matrix_from_sce(sce.example)
}
