% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{signature_matrix_from_singleCellExperiment}
\alias{signature_matrix_from_singleCellExperiment}
\title{signature_matrix_from_singleCellExperiment}
\usage{
signature_matrix_from_singleCellExperiment(
  singleCellExperiment,
  cellTypeVariable = "celltype",
  summary.method = "mean",
  assayName = "counts"
)
}
\arguments{
\item{singleCellExperiment}{An object of type \linkS4class{SingleCellExperiment}.}

\item{cellTypeVariable}{Variable name for cell type labels in \code{singleCellExperiment} 
coldata (e.g. "type1", "type2", etc.).}

\item{summary.method}{Summary statistic function to use.}

\item{assayName}{Name of expression matrix in \code{singleCellExperiment} assays (e.g. 
"counts").}
}
\value{
New Z signature matrix.
}
\description{
Calculate a Z signature matrix from object of type 
\linkS4class{SingleCellExperiment}.
}
\details{
Calculate a Z signature matrix from object of type 
\linkS4class{SingleCellExperiment}.
}
\examples{
singleCellExperiment.example <- random_singleCellExperiment()
signature_matrix_from_singleCellExperiment(singleCellExperiment.example)

}
