% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcheffects_simulation-utilities.R
\name{simulate_sce_donor_bias}
\alias{simulate_sce_donor_bias}
\title{simulate_sce_donor_bias}
\usage{
simulate_sce_donor_bias(donor.coeff.vector = NULL, ...)
}
\arguments{
\item{donor.coeff.vector}{Vector of donor bias values. If NULL, uses 
`rnorm(10, mean = 200, sd = 30)`.}

\item{...}{Additional arguments passed to `get_sce_donor_bias()` (see 
`?get_sce_donor_bias` for details).}
}
\value{
List containing the sce and ggplot objects.
}
\description{
Simulate and plot donor bias from a SingleCellExperiment.
}
\details{
Simulates a new SingleCellExperiment object with included donor bias,
and plots the means and variances on log scale to show dispersion biases 
across simulated donors.
}
\examples{
# simulate donor bias cell data, with default settings
# ld <- simulate_sce_donor_bias()

# dimensions of final sce object
# dim(ld[["sce"]]) 

# show the resulting smooth plot
# ld[["ggsmoooth"]] # overlay
# ld[["ggsmooth"]] + facet_wrap(~donor.coeff) # panels

}
