% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolutionParam-class.R
\name{[[,deconvolutionParam,ANY,ANY-method}
\alias{[[,deconvolutionParam,ANY,ANY-method}
\title{Inspect slot in \linkS4class{deconvolutionParam} object}
\usage{
\S4method{[[}{deconvolutionParam,ANY,ANY}(x, i)
}
\arguments{
\item{x}{Object to access.}

\item{i}{Slot to access.}
}
\value{
Contents of specified slot.

Object slot contents.
}
\description{
Inspect slot in \linkS4class{deconvolutionParam} object
}
\details{
Inspect slot in \linkS4class{deconvolutionParam} object
}
\examples{
example.data <- get_decon_example_data()

}
