% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typemarkersParam-class.R
\name{[[,typemarkersParam,ANY,ANY-method}
\alias{[[,typemarkersParam,ANY,ANY-method}
\title{Inspect slot in \linkS4class{typemarkersParam} object}
\usage{
\S4method{[[}{typemarkersParam,ANY,ANY}(x, i)
}
\arguments{
\item{x}{Object to access.}

\item{i}{Slot to access.}
}
\value{
Contents of specified slot.
}
\description{
Inspect slot in \linkS4class{typemarkersParam} object
}
\details{
Inspect slot in \linkS4class{typemarkersParam} object
}
