% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_simulation-utilities.R
\name{subsample_plots}
\alias{subsample_plots}
\title{subsample_plots}
\usage{
subsample_plots(results.table)
}
\arguments{
\item{results.table}{Either a table or a list of such tables, where the list
names correspond to experiment labels to use.}
}
\value{
List of ggplot objects
}
\description{
Plot results of subsample experiments.
}
\examples{
path <- file.path("inst","examples","results-table.csv")
path <- system.file("./inst/examples/results-table.csv", package = 'lute')
results.table <- read.csv(path)

# plot single results table
lgg <- subsample_plots(results.table)

# plot multiple results tables
lgg <- list("expt1" = results.table, "expt2" = results.table)

}
