% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_methods.R
\name{supported_marker_methods}
\alias{supported_marker_methods}
\title{supported_marker_methods}
\usage{
supported_marker_methods(
  varname = "marker_method_name",
  fname = "lute_marker-method_transfer_learning",
  verbose = FALSE
)
}
\arguments{
\item{varname}{Variable/column in lute_transfer_learning.rda dataset with 
strict deconvolution method names.}

\item{fname}{Name of the file containing supported deconvolution methods 
info.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
Vector of supported marker selection methods.
}
\description{
Show supported marker identification methods. These are load from the 
provided table in the lute package.
}
\examples{
supported_marker_methods()
}
\seealso{
supported_strict_methods
}
