% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{supported_strict_methods}
\alias{supported_strict_methods}
\title{supported_strict_methods}
\usage{
supported_strict_methods(
  varname = "strict_deconvolution_method_used",
  fname = "lute_transfer_learning"
)
}
\arguments{
\item{varname}{Variable/column in lute_transfer_learning.rda dataset with 
strict deconvolution method names.}

\item{fname}{Name of the file containing supported deconvolution methods 
info.}

\item{verbose}{Whether to show verbose status updates.}
}
\value{
List of supported strict deconvolution function names and 
descriptions.
}
\description{
Show supported strict deconvolution function. That is, functions which take
some reference/signature matrix Z and some bulk/convoluted signals matrix Y
and return the vector of predictions or proportions, p.
}
