% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanratiosParam-class.R
\name{typemarkers,meanratiosParam-method}
\alias{typemarkers,meanratiosParam-method}
\title{Cell type markers method for meanratiosParam}
\usage{
\S4method{typemarkers}{meanratiosParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{meanratiosParam}.}
}
\value{
Either a vector of gene markers, or a list containing such a vector 
with the original method outputs.
}
\description{
Defines the typemarkers method for \linkS4class{meanratiosParam}.
}
\details{
Takes an object of class \linkS4class{meanratiosParam} as input, 
returning either a vector of cell type gene markers, or (if 
\code{return.info == TRUE}) a list containing such a vector along with 
original function outputs.
}
