% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typemarkersParam-class.R
\name{typemarkers,typemarkersParam-method}
\alias{typemarkers,typemarkersParam-method}
\title{Method for class \linkS4class{typemarkersParam}}
\usage{
\S4method{typemarkers}{typemarkersParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{typemarkersParam}.}
}
\value{
Info related to gene markers for cell types.
}
\description{
Method for class \linkS4class{typemarkersParam}
}
