% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_generics.R
\name{typemarkers}
\alias{typemarkers}
\alias{Typemarkers,}
\alias{TypeMarkers}
\title{typemarkers}
\usage{
typemarkers(object)
}
\arguments{
\item{object}{A \linkS4class{typemarkersParam}-type object (see 
\code{?typemarkersParam}).}
}
\value{
By default, return a vector of marker genes.

If \code{return.info == TRUE}, provides detailed results, including original 
outputs.
}
\description{
Get cell type gene markers using standard accessors to supported functions.
}
\details{
This generic manages tasks for marker gene identification. In particular, it
takes a specified amount of marker genes to return per type.
}
\examples{
exampleList <- getDeconvolutionExampleData()


}
\seealso{
\linkS4class{typemarkersParam}
}
\author{
Sean Maden
}
