% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typemarkersParam-class.R
\docType{class}
\name{typemarkersParam-class}
\alias{typemarkersParam-class}
\alias{TypemarkersParam-class,}
\alias{TypeMarkersParam-class}
\title{typemarkersParam-class}
\arguments{
\item{markers.per.type}{Number of top markers to get per cell type.}

\item{return.info}{Whether to return metadata and original method outputs 
with predicted proportions.}
}
\description{
Main constructor for class to manage mappings to the \code{typemarkers()} 
generic.
}
\details{
Main constructor for class \linkS4class{typemarkersParam}.
}
\examples{
lexample <- get_decon_example_data()

}
\seealso{
meanratiosParam
}
