% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typemarkersParam-class.R
\name{typemarkersParam}
\alias{typemarkersParam}
\title{Make new object of class typemarkersParam}
\usage{
typemarkersParam(markersPerType = 20, returnInfo = FALSE)
}
\arguments{
\item{markersPerType}{Bulk mixed signals matrix of samples, which can be 
matched to single-cell samples.}

\item{returnInfo}{Whether to return metadata and original marker selection 
method outputs with predicted proportions.}
}
\value{
New object of class \linkS4class{typemarkersParam}.
}
\description{
Main constructor for class \linkS4class{typemarkersParam}.
}
\details{
This is the main parent class for cell type gene marker 
identification methods. Currently supported methods and their child classes include:

1. Mean Ratios: The method DeconvoBuddies::get_mean_ratios2(), supported by the
class meanratiosParam.
}
\examples{
example.data <- get_decon_example_data()

}
