% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{ypb_from_sce}
\alias{ypb_from_sce}
\title{ypb_from_sce}
\usage{
ypb_from_sce(sce, assay.name, celltype.variable, S = NULL)
}
\arguments{
\item{sce}{An object of type \linkS4class{SingleCellExperiment}.}

\item{assay.name}{Name of expression matrix in \code{sce} assays.}

\item{celltype.variable}{Variable name for cell type labels in \code{sce} 
coldata.}

\item{S}{Vector of cell type size scale factors. Optional.}
}
\value{
Matrix of simulated bulk convoluted signals.
}
\description{
Get pseudobulk from a SingleCellExperiment object.
}
