% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{ypb_from_singleCellExperiment}
\alias{ypb_from_singleCellExperiment}
\title{ypb_from_singleCellExperiment}
\usage{
ypb_from_singleCellExperiment(
  singleCellExperiment,
  assayName = "counts",
  cellTypeVariable = "celltype",
  sample.id.variable = NULL,
  S = NULL
)
}
\arguments{
\item{singleCellExperiment}{An object of type \linkS4class{SingleCellExperiment}.}

\item{assayName}{Name of expression matrix in \code{singleCellExperiment} assays.}

\item{cellTypeVariable}{Variable name for cell type labels in \code{singleCellExperiment} 
coldata.}

\item{sample.id.variable}{Variable name for sample/group ID labels in 
\code{singleCellExperiment} coldata.}

\item{S}{Vector of cell type size scale factors. Optional.}
}
\value{
Matrix of simulated bulk convoluted signals.
}
\description{
Get pseudobulk from a SingleCellExperiment object.
}
\examples{
singleCellExperiment.example <- random_singleCellExperiment()
ypb_from_singleCellExperiment(singleCellExperiment.example)

}
