% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcheffects_simulation-utilities.R
\name{ypb_fromtypes}
\alias{ypb_fromtypes}
\title{ypb_fromtypes}
\usage{
ypb_fromtypes(Z, P)
}
\arguments{
\item{Z}{Signature matrix containing pure type signals (rows = markers, 
columns = types).}

\item{P}{Vector of true proportions.}
}
\value{
Ypb, a new pseudobulked sample of convoluted signals.
}
\description{
Makes a pseuobulked convoluted signals sample from types data. Uses the 
formula:
}
\details{
Y = P * Z
}
\examples{
P <- c(0.75, 0.25)
Z <- matrix(c(0, 1, 0, 1), nrow = 2)
Ypb <- ypb_fromtypes(Z, P)
}
