% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{z_matrix_from_sce}
\alias{z_matrix_from_sce}
\title{z_matrix_from_sce}
\usage{
z_matrix_from_sce(
  singleCellExperiment,
  cellTypeVariable = "celltype",
  summaryMethod = "mean",
  assayName = "counts"
)
}
\arguments{
\item{singleCellExperiment}{An object of type \linkS4class{SingleCellExperiment}.}

\item{cellTypeVariable}{Variable name for cell type labels in \code{singleCellExperiment} 
coldata (e.g. "type1", "type2", etc.).}

\item{summaryMethod}{Summary statistic function to use.}

\item{assayName}{Name of expression matrix in \code{singleCellExperiment} assays (e.g. 
"counts").}
}
\value{
New Z signature matrix.
}
\description{
Calculate a Z signature matrix (referenceExpression) from object of type 
\linkS4class{SingleCellExperiment}.
}
\details{
Calculate a Z signature matrix from object of type 
\linkS4class{SingleCellExperiment}.
}
\examples{
singleCellExperiment.example <- randomSingleCellExperiment()
z_matrix_from_sce(singleCellExperiment.example)

}
