% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexpr_methods.R
\name{analyze_anova}
\alias{analyze_anova}
\title{analyze_anova}
\usage{
analyze_anova(
  sce,
  ngene.sample = 2000,
  model = "expr ~ celltype * donor",
  return.var = c("sumsq", "perc.var"),
  type = "complex"
)
}
\arguments{
\item{ngene.sample}{Number of genes to sample at random.}

\item{model}{Character string of the model to use. By default, provides an
interaction between donor and cell type, and the response/dependent variable
is the individual gene.}

\item{seed.num}{Random seed to set for reproducibility.}
}
\description{
Main function to perform ANOVA analyses on a series of assays, such as from
a SingleCellExperiment or SummarizedExperiment object.
}
\examples{
sce <- random_sce()
}
