% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{biasexpt}
\alias{biasexpt}
\title{biasexpt}
\usage{
biasexpt(
  df,
  Ypb,
  P,
  donor.unadj = NULL,
  donor.adj.method = "combat",
  plot.biasadj = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{df}{Donor data.frame.}

\item{Ypb}{Pseudobulk sample data.}

\item{P}{Vector of true proportions, where each entry corresponds to a type.}

\item{donor.unadj}{Vector of unadjusted donor signals.}

\item{donor.adj.method}{Method to adjust for donor bias. Can be either 
"limma", "var_denom", "sd_denom", or NULL. If NULL, skip this step.}

\item{plot.biasadj}{Whether to make scatterplot of donor summary signals
before and after bias adjustment.}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Additional arguments passed to `donoradj()`.}
}
\value{
lexpt, experiment results list.
}
\description{
Run a single donor bias experiment.
}
\examples{
# simulate donor signals data
donordf <- rand_donor_marker_table()
# get ypb
P <- c(0.75, 0.25)
ktotal <- length(P)
Z <- matrix(df[,"donor1"], ncol = ktotal)
Ypb <- ypb_fromtypes(Z = Z, P = P)
# get bias expt results
li <- biasexpt(df = donordf, Ypb = Ypb)
}
