% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisqueParam-class.R
\docType{class}
\name{bisqueParam-class}
\alias{bisqueParam-class}
\alias{BisqueParam-class}
\title{bisqueParam-class}
\description{
Applies the BisqueRNA::ReferenceBasedDecomposition() implementation of the 
Bisque deconvolution algorithm.
}
\details{
Main constructor for class \linkS4class{bisqueParam}.
}
\examples{
# get data
lexample <- lute:::.get_decon_example_data_bisque()
sc.eset <- lexample[["sc.eset"]]
y.eset <- lexample[["y.eset"]]
# example params
batch.variable <- "SubjectName"
celltype.variable <- "cellType"

# get param object
param <- bisqueParam(y.eset = y.eset, sc.eset = sc.eset,
                     batch.variable = "SubjectName",
                     celltype.variable = "cellType")

# get just predictions
res <- deconvolution(param)

# get full results
param@return.info <- TRUE
res <- deconvolution(param)
}
\seealso{
\linkS4class{deconParam}, \linkS4class{referencebasedParam}, \linkS4class{independentbulkParam}
}
