% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisqueParam-class.R
\name{bisqueParam}
\alias{bisqueParam}
\title{Make new object of class bisqueParam}
\usage{
bisqueParam(
  y = NULL,
  yi = NULL,
  z = NULL,
  s = NULL,
  y.eset = NULL,
  sc.data = NULL,
  assay.name = "counts",
  batch.variable = "batch.id",
  celltype.variable = "celltype",
  use.overlap = FALSE,
  return.info = FALSE
)
}
\arguments{
\item{y}{Bulk mixed signals matrix of samples, which can be matched to single-cell samples.}

\item{yi}{Bulk mixed signals matrix of independent samples, which should not overlap samples in y.}

\item{z}{Signature matrix of cell type-specific signals. If not provided, can be computed from a
provided ExpressionSet containing single-cell data.}

\item{s}{Cell size factor transformations of length equal to the K cell types to deconvolve.}

\item{y.eset}{ExpressionSet of bulk mixed signals.}

\item{sc.data}{SummarizedExperiment-type object of single-cell transcriptomics data. Accepts
ExpressionSet, SummarizedExperiment, and SingleCellExperiment object types.}

\item{assay.name}{Expression data type (e.g. counts, logcounts, tpm, etc.).}

\item{batch.variable}{Name of variable identifying the batches in sc.eset pData/coldata.}

\item{celltype.variable}{Name of cell type labels variable in sc.eset pData/coldata.}

\item{use.overlap}{Whether to deconvolve samples overlapping bulk and sc 
esets (logical, FALSE).}

\item{return.info}{Whether to return metadata and original method outputs with predicted proportions.}
}
\value{
New object of class \linkS4class{bisqueParam}.
}
\description{
Main constructor for class \linkS4class{bisqueParam}.
}
\details{
Takes standard inputs for the Bisque method. If user provides matrices, will convert these
into ExpressionSet objects compatible with the main bisque method.
}
\examples{
## get data
lexample <- get_decon_example_data_bisque()
input_y_eset <- lexample[["y.eset"]][,seq(10)]
input_yi <- exprs(lexample[["y.eset"]])
input_yi <- input_yi[,c(11:ncol(input_yi))]

## get param object
param <- bisqueParam(y.eset=input_y_eset, yi=input_yi,
                     sc.data=lexample[["sc.eset"]], 
                     batch.variable="SubjectName", 
                     celltype.variable="cellType", 
                     use.overlap=FALSE)

## get predicted proportions
res <- deconvolution(param)

}
