% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisqueParam-class.R
\name{bisqueParam}
\alias{bisqueParam}
\title{Make new object of class bisqueParam}
\usage{
bisqueParam(
  y = NULL,
  yi = NULL,
  z = NULL,
  s = NULL,
  y.eset = NULL,
  sc.eset = NULL,
  batch.variable = "batch.id",
  celltype.variable = "celltype",
  return.info = FALSE
)
}
\arguments{
\item{y}{Bulk mixed signals matrix of samples, which can be matched to single-cell samples.}

\item{yi}{Bulk mixed signals matrix of independent samples, which should not overlap samples in y.}

\item{z}{Signature matrix of cell type-specific signals. If not provided, can be computed from a
provided ExpressionSet containing single-cell data.}

\item{s}{Cell size factor transformations of length equal to the K cell types to deconvolve.}

\item{y.eset}{ExpressionSet of bulk mixed signals.}

\item{sc.eset}{ExpressionSet of single-cell transcriptomics data.}

\item{batch.variable}{Name of variable identifying the batches in sc.eset pData/coldata.}

\item{celltype.variable}{Name of cell type labels variable in sc.eset pData/coldata.}

\item{return.info}{Whether to return metadata and original method outputs with predicted proportions.}
}
\description{
Main constructor for class \linkS4class{bisqueParam}.
}
\details{
Takes standard inputs for the Bisque method. If user provides matrices, will convert these
into ExpressionSet objects compatible with the main bisque method.
}
