% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_classes.R
\docType{class}
\name{cellProportionsPredictions-class}
\alias{cellProportionsPredictions-class}
\title{cellProportionsPredictions-class}
\arguments{
\item{predictions.table}{Table containing cell type predictions.}

\item{cell.type.vector}{Character vector of cell type labels.}

\item{sample.id.vector}{Character vector of sample id labels.}
}
\description{
Class for cell type predictions.
}
\details{
Main constructor for class \linkS4class{cellProportionsPredictions}.
}
\examples{
new("cellProportionsPredictions")
ptable <- matrix(sample(100,50),nrow=10)
colnames(ptable) <- paste0("cell_type",seq(ncol(ptable)))
rownames(ptable) <- paste0("sample", seq(nrow(ptable)))
pred <- cellProportionsPredictions(ptable)
pred
}
