% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_classes.R
\name{cellProportionsPredictions}
\alias{cellProportionsPredictions}
\title{Make new cellProportionsPredictions object.}
\usage{
cellProportionsPredictions(
  predictionsTable,
  cellTypeVector = NULL,
  sampleIdVector = NULL
)
}
\arguments{
\item{predictionsTable}{Table of cell type predictions.}

\item{cellTypeVector}{Character vector of cell type labels.}

\item{sampleIdVector}{Character vector of sample id labels.}
}
\value{
New cellProportionsPredictions object.

New cellProportionsPredictions object.
}
\description{
Make new cellProportionsPredictions object.
}
\examples{
exampleData <- get_decon_example_data()

}
