% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_classes.R
\name{cellProportionsPredictions}
\alias{cellProportionsPredictions}
\title{Make new cellProportionsPredictions object.}
\usage{
cellProportionsPredictions(
  predictions.table,
  cell.type.vector = NULL,
  sample.id.vector = NULL
)
}
\arguments{
\item{predictions.table}{Table of cell type predictions.}

\item{cell.type.vector}{Character vector of cell type labels.}

\item{sample.id.vector}{Character vector of sample id labels.}
}
\value{
New cellProportionsPredictions object.

New cellProportionsPredictions object.
}
\description{
Make new cellProportionsPredictions object.
}
\examples{
example.data <- get_decon_example_data()

}
