% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{check_donordf}
\alias{check_donordf}
\title{check_donordf}
\usage{
check_donordf(df)
}
\arguments{
\item{df}{Data.frame to check.}
}
\value{
boolean, TRUE if df is valid, FALSE otherwise
}
\description{
Checks whether data.frame is a valid simulated donor signals data.frame by
evaluating the column names.
}
