% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{decon_analysis}
\alias{decon_analysis}
\title{decon_analysis}
\usage{
decon_analysis(lgv, lpv, lsv, verbose = FALSE, sce = NULL, ...)
}
\arguments{
\item{lgv}{List of marker expression for reference/signature matrix Z. If 
length(lgv) > length(lpv), only use up to the number of iterations in lpv.}

\item{lpv}{List of type proportions to make Y and compare p predictions. The
length of this list is considered the target number of simulation iterations 
for the run.}

\item{lsv}{List of size factor values. If length(lsv) > length(lpv), only use
up to the number of iterations in lpv. If NULL, S-transformation experiments
aren't performed, and Y is calculated as $P*Z$ rather than $P*ZS$.}

\item{verbose}{Whether to show verbose status updates.}

\item{sce}{SingleCellExperiment or SummarizedExperiment object.}

\item{...}{Additional arguments passed to `kexpr_sce()`.}
}
\value{
return
}
\description{
Do simulations, get results df and plots.
}
\examples{
# example:
}
\seealso{
decon_results,
}
