% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R, R/sim_utils_sce.R
\name{decon_analysis}
\alias{decon_analysis}
\title{decon_analysis}
\usage{
decon_analysis(lgv, lpv, lsv, verbose = FALSE)

decon_analysis(lgv, lpv, lsv, verbose = FALSE)
}
\arguments{
\item{lgv}{List of marker expression for reference/signature matrix Z.}

\item{lpv}{List of type proportions to make Y and compare p predictions.}

\item{lsv}{List of size factor values.}

\item{verbose}{Whether to show verbose status updates.}
}
\description{
Do simulations, get results df and plots.

Do simulations, get results df and plots.
}
\examples{

# example:


# example:

}
\seealso{
decon_results,

decon_results,
}
