% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconrnaseqParam-class.R
\docType{class}
\name{deconrnaseqParam-class}
\alias{deconrnaseqParam-class}
\alias{DeconRNASeqParam-class}
\title{deconrnaseqParam-class}
\description{
Main constructor for class to manage mappings to the deconvolution
method function \code{\link{DeconRNASeq::DeconRNASeq()}}.
}
\details{
Main constructor for class \linkS4class{deconrnaseqParam}.
}
\examples{
# example
lexample <- .get_decon_example_data()
param <- musicParam(s = lexample[["s"]], y = lexample[["y"]], 
z = lexample[["z"]])

# return only predicted proportions
deconvolution(param)
# type1     type2 
# 0.9819837 0.0180163

# return full results
param@return.info <- T
names(deconvolution(param))
# [1] "predictions" "result.info" "metadata"

}
\seealso{
\linkS4class{deconParam}
}
