% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconrnaseqParam-class.R
\docType{class}
\name{deconrnaseqParam-class}
\alias{deconrnaseqParam-class}
\alias{DeconRNASeqParam-class}
\title{deconrnaseqParam-class}
\description{
Main constructor for class to manage mappings to the deconvolution
method function \code{DeconRNASeq::DeconRNASeq()}.
}
\details{
Main constructor for class \linkS4class{deconrnaseqParam}.
}
\examples{
lexample <- lute:::.get_decon_example_data() # get example data 
param <- deconrnaseqParam(s = lexample[["s"]], y = lexample[["y"]], 
z = lexample[["z"]])

# return only predicted proportions
deconvolution(param)

}
\seealso{
\linkS4class{deconParam}
}
