% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisqueParam-class.R
\name{deconvolution,bisqueParam-method}
\alias{deconvolution,bisqueParam-method}
\title{Deconvolution method for bisqueParam}
\usage{
\S4method{deconvolution}{bisqueParam}(object)
}
\arguments{
\item{object}{Object of type \linkS4class{bisqueParam} (see 
\code{?bisqueParam}).}
}
\value{
Either a vector of predicted proportions, or a list containing 
predictions, metadata, and original outputs.
}
\description{
Main method to access the Bisque deconvolution method from the main lute 
\code{deconvolution} generic.
}
\details{
Takes an object of class \linkS4class{bisqueParam} as input, 
returning a list.
}
\examples{
## get data
exampleList <- getDeconvolutionExampleData()
bulkExpressionSet <- exampleList[["bulkExpressionSet"]][,seq(10)]
bulkExpression <- exprs(exampleList[["bulkExpressionSet"]])
bulkExpression <- bulkExpression[,c(11:ncol(bulkExpression))]

## get param object
param <- bisqueParam(bulkExpressionSet=bulkExpressionSet, 
                     bulkExpressionIndependent=bulkExpression,
                     scData=exampleList[["singleCellExpressionSet"]], 
                     batchVariable="SubjectName", 
                     cellTypeVariable="cellType", 
                     useOverlap=FALSE)

## get predicted proportions
res <- deconvolution(param)

}
\references{
Brandon Jew and Marcus Alvarez (2021). BisqueRNA: Decomposition 
of Bulk Expression with Single-Cell Sequencing. CRAN, R package version 
1.0.5. URL: https://CRAN.R-project.org/package=BisqueRNA

Brandon Jew et al. Accurate estimation of cell composition in bulk 
expression through robust integration of single-cell information. 
Nat Commun 11, 1971 (2020). https://doi.org/10.1038/s41467-020-15816-6
}
