% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconrnaseqParam-class.R
\name{deconvolution,deconrnaseqParam-method}
\alias{deconvolution,deconrnaseqParam-method}
\title{Method for deconParam}
\usage{
\S4method{deconvolution}{deconrnaseqParam}(object)
}
\description{
Method for deconParam
}
