% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentbulkParam-class.R
\name{deconvolution,independentbulkParam-method}
\alias{deconvolution,independentbulkParam-method}
\title{Deconvolution method for class \linkS4class{independentbulkParam}}
\usage{
\S4method{deconvolution}{independentbulkParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{independentbulkParam}.}
}
\description{
Function to perform standard operations prior to deconvolution (a.k.a. 
"deconvolution prep") for an object of class 
\linkS4class{independentbulkParam}.
}
\details{
Takes an object of \linkS4class{independentbulkParam} class as 
input, and returns a list with the filtered/checked/parsed experiment objects.
}
\examples{
new("independentbulkParam")

}
