% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music2Param-class.R
\name{deconvolution,music2Param-method}
\alias{deconvolution,music2Param-method}
\title{Deconvolution method for music2Param}
\usage{
\S4method{deconvolution}{music2Param}(object)
}
\value{
Either a vector of predicted proportions, or a list containing predictions, metadata, 
and original outputs.
}
\description{
Main method to access the MuSiC2 deconvolution method from the main lute deconvolution genetic.
}
\details{
Takes an object of class music2Param as input, returning a list or vector of predicted 
cell type proportions.
}
