% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencebasedParam-class.R
\name{deconvolution,referencebasedParam-method}
\alias{deconvolution,referencebasedParam-method}
\title{Deconvolution generic behavior for object of class \linkS4class{referencebasedParam}}
\usage{
\S4method{deconvolution}{referencebasedParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{referencebasedParam}.}
}
\description{
Deconvolution generic behavior for object of class \linkS4class{referencebasedParam}
}
\details{
Method for behavior of deconvolution generic when called for object of class 
\linkS4class{referencebasedParam}
}
