% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdcParam-class.R
\name{deconvolution,scdcParam-method}
\alias{deconvolution,scdcParam-method}
\title{Deconvolution method for scdcParam}
\usage{
\S4method{deconvolution}{scdcParam}(object)
}
\value{
Either a vector of predicted proportions, or a list containing 
the predictions, metadata, and original outputs.
}
\description{
Main method to access the SCDC deconvolution method from the main lute deconvolution genetic.
}
\details{
Takes an object of class scdcParam as input, returning a list or vector of predicted 
cell type proportions.
}
\references{
Meichen Dong. SCDC: Bulk RNA-seq deconvolution by scRNA-seq with multi-reference 
datasets. (2023) GitHub, R package version 0.0.0.9000. URL: https://github.com/meichendong/SCDC

Meichen Dong, Aatish Thennavan, Eugene Urrutia, Yun Li, Charles M Perou, Fei 
Zou, Yuchao Jiang, SCDC: bulk gene expression deconvolution by multiple 
single-cell RNA sequencing references, Briefings in Bioinformatics, Volume 
22, Issue 1, January 2021, Pages 416–427, 
https://doi-org.proxy1.library.jhu.edu/10.1093/bib/bbz166
}
