% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_experiment.R
\name{deconvolution.results.plots.permutations}
\alias{deconvolution.results.plots.permutations}
\title{deconvolution.results.plots.permutations}
\usage{
deconvolution.results.plots.permutations(results.table, verbose = FALSE)
}
\arguments{
\item{results.table}{Table of permutation experiment results.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
List containing experiment results plots.
}
\description{
Permutation experiment plots.
}
\details{
Plot functions for permutation experiments. The \code{results.table}
should contain columns for "group.id.signature", for the signature matrix id, 
and "sample.id", for the pseudobulk sample id.
}
\examples{
sce.example <- random_sce(num.genes = 100, num.cells = 100)
s <- c("type1" = 3, "type2" = 10)
sce.example[["sample.id"]] <- c(rep("sample1", 10), rep("sample2", 50), rep("sample1", 40))
experiment <- deconvolution.experiment.permute.groups(sce = sce.example, s = s)
experiment$plots.list$`miss-matched`$type1$proportions.scatterplot

}
\seealso{
\code{deconvolution.experiment}, \code{plot.deconvolution.results}
}
