% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_simulation-utilities.R
\name{dfres_k}
\alias{dfres_k}
\title{dfres_k}
\usage{
dfres_k(dfres, verbose = FALSE)
}
\arguments{
\item{dfres}{Results data.frame containing simulation outcomes.}

\item{verbose}{Whether to show verbose status updates.}
}
\value{
dfk, data.frame of k/type-wise results summaries.
}
\description{
Make k-wise results data.frame from simulation outcomes.
}
