% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{donor_marker_biasexpt}
\alias{donor_marker_biasexpt}
\title{donor_marker_biasexpt}
\usage{
donor_marker_biasexpt(
  offsetv = c(1, 10),
  P = c(0.25, 0.75),
  donor.adj.method = NULL,
  plot.biasadj = TRUE,
  cname.donorsummary = "donor.combn.all.mean",
  gindexv = c(1, 2),
  ndonor = 10,
  seed.num = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{offsetv}{Vector of donor offset values. This is used to define the
test groups for donor bias experiments. Equal offsets are applied to positive
and negative marker signal distributions.}

\item{P}{Vector of true proportions.}

\item{donor.adj.method}{Method to adjust for donor bias. Can be either 
"limma", "var_denom", "sd_denom", or NULL. If NULL, skip this step.}

\item{plot.biasadj}{Whether to make scatterplot of donor summary signals
before and after bias adjustment.}

\item{cname.donorsummary}{Name of column containing the donor summary data
with which to perform experiment.}

\item{gindexv}{Vector of type indices for the G markers. See `?random_lgv` 
for details.}

\item{ndonor}{Total number of donors to simulate.}

\item{seed.num}{Seed value for random sizes in lsv, in case lsv is NULL.}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Arguments passed to function `donoradj()`.}
}
\value{
List of experiment results and experiment objects.
}
\description{
Compare predictions with and without donor bias corrections. Parses various
methods for performing the donor bias adjustment, returning predicted 
type proportions for adjusted and unadjusted signature matrices.
}
\examples{
lb <- donor_marker_biasexpt()
}
