% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcheffects_simulation-utilities.R
\name{donoradj_combat}
\alias{donoradj_combat}
\title{donoradj_combat}
\usage{
donoradj_combat(df, return.type = "donor.adj", verbose = FALSE)
}
\arguments{
\item{df}{Data.frame containing marker signals and pheno data.}

\item{return.type}{Type of data to return. Accepts either "donor.adj" for a 
vector of adjusted donor signals, "mexpr" for a matrix of adjusted 
signals, or "donordf" for a donordf-type data.frame.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
madj, matrix of adjusted marker signals.
}
\description{
Use method sva::ComBat() to adjust for donor bias.
}
\examples{
df <- random_donordf()
donoradj_combat(df)
}
