% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donordf_methods.R
\name{donordf_from_mexpr}
\alias{donordf_from_mexpr}
\title{donordf_from_mexpr}
\usage{
donordf_from_mexpr(
  mexpr,
  groupv.cd,
  typev.cd,
  typev.rd = NULL,
  verbose = FALSE
)
}
\arguments{
\item{mexpr}{An expression matrix (rows = markers/genes, columns = 
samples/type data). Column names correspond to the types used.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
New table of type `donor.data.frame`.
}
\description{
Makes a valid donordf object from an expression matrix (e.g. rows = 
markers/genes, columns = samples/type data)
}
\examples{
sce = random_sce()
groupv.cd = c(rep("donor1", 7), rep("donor2", 3))
typev.cd = sce[["celltype"]]
mexpr <- assays(sce)$counts
df <- donordf_from_mexpr(mexpr = mexpr, groupv.cd = groupv.cd, 
typev.cd = typev.cd)

}
