% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicParam-class.R
\docType{class}
\name{epicParam-class}
\alias{epicParam-class}
\title{Run EPIC}
\description{
Runs the EPIC::EPIC() deconvolution algorithm.
}
\examples{
# example
lexample <- .get_decon_example_data()
param <- epicParam(s = lexample[["s"]], y = lexample[["y"]], 
z = lexample[["z"]])

# return only predicted proportions
deconvolution(param)

# return full results
param@return.info <- T
names(deconvolution(param))
# [1] "predictions" "result.info" "metadata"

}
