% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_conversions.R
\name{eset_to_sce}
\alias{eset_to_sce}
\title{eset_to_sce
Convert ExpressionSet to SingleCellExperiment.}
\usage{
eset_to_sce(eset, assay.name = "counts")
}
\arguments{
\item{eset}{Object of type ExpressionSet (see \code{?ExpressionSet}).}

\item{assay.name}{Name of new assay in new SingleCellExperiment object.}
}
\value{
ExpressionSet.
}
\description{
eset_to_sce
Convert ExpressionSet to SingleCellExperiment.
}
\examples{
eset <- get_decon_example_data_bisque()$sc.eset
eset_to_sce(eset)

}
