% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_conversions.R
\name{eset_to_se}
\alias{eset_to_se}
\title{eset_to_se}
\usage{
eset_to_se(expressionSet, assayName = "counts")
}
\arguments{
\item{expressionSet}{Object of type ExpressionSet (see \code{?ExpressionSet}).}

\item{assayName}{Name of assay to store in new SummarizedExperiment object.}
}
\value{
New object of type SummarizedExperiment.
}
\description{
Convert ExpressionSet to SummarizedExperiment.
}
\examples{
expressionSet <- get_decon_example_data_bisque()$singleCellExpressionSet
eset_to_se(expressionSet "counts")

}
