% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon_workflow_functions.R
\name{filter_value_cells}
\alias{filter_value_cells}
\title{filter_value_cells}
\usage{
filter_value_cells(
  sce,
  filter.term = c("zerocount", "NA"),
  remove.cells = TRUE,
  max.value.freq = 0.25,
  assayname = "counts",
  append.metadata = TRUE,
  new.metadata.name = NULL,
  verbose = FALSE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{filter.term}{Term for filter. Either "zerocount" or "NA".}

\item{remove.cells}{Whether to remove cells exceeding NA filter from sce 
prior to returning.}

\item{max.value.freq}{The upper threshold for tolerated NA frequency. Cells with 
NA frequency exceeding this are flagged for removal.}

\item{assayname}{Name of assay in assays(sce) to check for NAs.}

\item{append.metadata}{Whether to append metadata summary of preprocessing 
parameters and results.}

\item{new.metadata.name}{Name of new metadata object to append (see 
append.metadata argument).}

\item{verbose}{Whether to include verbose status messages.}
}
\value{
scef, filtered SingleCellExperiment with optional appended 
preprocessing metadata.
}
\description{
Filter cells on NA frequency.
}
\examples{
set.seed(0)
sce <- random_sce(zero.include = T, zero.fract = 0.3)
scef <- filter_value_cells(sce, filter.term = "zerocount", verbose = T,
                          max.value.freq = 0.1)
metadata(scef)$filter.zerocount.by.cell$df.cell
}
