% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmarkersParam-class.R
\docType{class}
\name{findmarkersParam-class}
\alias{findmarkersParam-class}
\alias{FindmarkersParam-class,}
\alias{findMarkersParam-class}
\title{findmarkersParam-class}
\arguments{
\item{assay.name}{Name of expression matrix in sce assays.}

\item{sce}{Object of type SingleCellExperiment.}

\item{celltype.variable}{Name of cell type variable in sce coldata.}

\item{test.type}{Test type (see ?findMarkers for options).}
}
\value{
New object.
}
\description{
class definition for findmarkersParam, which uses scran::findMarkers()
}
\details{
Main constructor for class \linkS4class{findmarkersParam}.
}
\examples{
lexample <- get_decon_example_data()
sce.example <- random_sce()
new.param <- findmarkersParam(sce = sce.example, 
celltype.variable = "celltype", markers.per.type = 5)
markers <- typemarkers(new.param)

}
\seealso{
\linkS4class{typemarkersParam}
}
