% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmarkersParam-class.R
\name{findmarkersParam}
\alias{findmarkersParam}
\title{Make new object of class findmarkersParam}
\usage{
findmarkersParam(
  sce,
  assay.name = "counts",
  celltype.variable = "cellType",
  test.type = "wilcox",
  markers.per.type = 20,
  return.info = FALSE
)
}
\arguments{
\item{sce}{Object of type SingleCellExperiment.}

\item{assay.name}{Name of expression matrix in sce assays.}

\item{celltype.variable}{Name of cell type variable in sce coldata.}

\item{test.type}{Test type (see ?findMarkers for options).}

\item{markers.per.type}{Number of top markers to get per cell type.}

\item{return.info}{Whether to return metadata and original method outputs 
with predicted proportions.}
}
\value{
Object of class \linkS4class{findmarkersParam}
}
\description{
Main constructor for class \linkS4class{findmarkersParam}.
}
\details{
Main class for mapping arguments to the findMarkers method 
implemented as \code{scran::findMarkers()}.
}
\seealso{
\linkS4class{typemarkersParam}
}
