% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{get_celltypes_from_sce}
\alias{get_celltypes_from_sce}
\title{get_celltypes_from_sce}
\usage{
get_celltypes_from_sce(sce, celltype.variable = "celltype")
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{celltype.variable}{Variable containing cell type labels.}
}
\value{
List of cell type variable metadata and values.
}
\description{
Extract cell type values from SingleCellExperiment.
}
\examples{
lexample <- get_decon_example_data()

}
