% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{get_decon_example_data}
\alias{get_decon_example_data}
\title{get_decon_example_data}
\usage{
get_decon_example_data(
  num.bulk.samples = 2,
  num.markers = 10,
  num.types = 2,
  seed.num = 0
)
}
\arguments{
\item{num.bulk.samples}{Bulk samples count.}

\item{num.markers}{Markers count.}

\item{num.types}{Cell types count.}

\item{seed.num}{Random seed.}
}
\value{
Example data as list.
}
\description{
Make example data for deconvolution.
}
