% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{get_decon_example_data_bisque}
\alias{get_decon_example_data_bisque}
\title{get_decon_example_data_bisque}
\usage{
get_decon_example_data_bisque(
  num.bulk.samples = 100,
  num.markers = 1000,
  num.cells = 1000,
  num.types = 2,
  seed.num = 0
)
}
\arguments{
\item{num.bulk.samples}{Bulk samples count.}

\item{num.markers}{Markers count.}

\item{num.cells}{Cells count.}

\item{num.types}{Cell types count.}

\item{seed.num}{Random seed.}
}
\value{
Example data as list.
}
\description{
Get example data for Bisque algorithm.
}
