% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_conversions.R
\name{get_eset_from_matrix}
\alias{get_eset_from_matrix}
\title{get_eset_from_matrix}
\usage{
get_eset_from_matrix(inputMatrix, batchVariable = "SampleName")
}
\arguments{
\item{inputMatrix}{User-specified expression matrix.}

\item{batchVariable}{Name of the batch variable.}
}
\value{
ExpressionSet.
}
\description{
Makes an ExpressionSet from a matrix.
}
\examples{
example.data <- get_decon_example_data()

}
