% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_methods.R
\name{get_groupstat_df}
\alias{get_groupstat_df}
\title{get_groupstat_df}
\usage{
get_groupstat_df(exprf, groupstat, summarytype)
}
\arguments{
\item{exprf}{Filtered expression matrix.}

\item{groupstat}{Vector of valid group statistics.}

\item{summarytype}{Type of summary. Either colData or rowData.}
}
\value{
Table of type data.frame containing summary statistics.
}
\description{
Get grouped statistics from an expression matrix.
}
