% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zref_utils.R
\name{get_types}
\alias{get_types}
\title{get_types}
\usage{
get_types(
  sce,
  top.markers = 20,
  bind.marker.results = TRUE,
  marker.method = "meanratio2",
  type.method = "mean",
  return.type = "SummarizedExperimentTypes",
  verbose = T,
  ...
)
}
\arguments{
\item{sce}{Either a SummarizedExperiment or SingleCellExperiment object.}

\item{bind.marker.results}{Whether to bind valid flat outputs from marker
searches to the colData in the returned results object.}

\item{marker.method}{Valid method to identify markers from sce object.}

\item{type.method}{Valid method to obtain type-level expression/signals.}

\item{return.type}{Class of results object to return.}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Additional arguments to pass for marker discovery method.}

\item{num.top.markers}{Number of top markers to retain by type (e.g. total 
markers is k*num.top.markers).}
}
\value{
Either a SummarizedExperimentTypes or or MultiAssayExperiment object.
}
\description{
Get type data from a SingleCellExperiment, returning either an object of 
either class SummarizedExperimentTypes or MultiAssayExperiment.
}
