% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentbulkParam-class.R
\docType{class}
\name{independentbulkParam-class}
\alias{independentbulkParam-class}
\title{independentbulkParam-class}
\arguments{
\item{bulkExpressionIndependent}{Bulk mixed signals matrix of independent 
samples, which should not overlap samples in y.}
}
\value{
New object.
}
\description{
Class and methods for managing methods requiring independent bulk samples.
}
\details{
The main purpose of this class is to compare bulk sample data 
between the passed objects y and yi. Since we assume yi contains the 
independent bulk samples, it should not have overlapping sample IDs 
(colnames), and it should have overlapping marker IDs (rownames) compared to 
the reference bulk samples y.
}
\examples{
new("independentbulkParam")

}
\seealso{
\linkS4class{deconParam}, \linkS4class{referencebasedParam}
}
