% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentbulkParam-class.R
\name{independentbulkParam}
\alias{independentbulkParam}
\title{Make a new independentbulkParam object}
\usage{
independentbulkParam(
  y = NULL,
  yi = NULL,
  z = NULL,
  s = NULL,
  return.info = FALSE
)
}
\description{
Function to make a new object of class independentbulkParam
}
