% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentbulkParam-class.R
\name{independentbulkParam}
\alias{independentbulkParam}
\title{Make a new \linkS4class{independentbulkParam} object}
\usage{
independentbulkParam(
  bulkExpression = NULL,
  bulkExpressionIndependent = NULL,
  referenceExpression = NULL,
  cellScaleFactors = NULL,
  returnInfo = FALSE
)
}
\arguments{
\item{bulkExpression}{Bulk mixed signals matrix of samples, which can be 
matched to single-cell samples.}

\item{bulkExpressionIndependent}{Bulk mixed signals matrix of independent 
samples, which should not overlap samples in y.}

\item{referenceExpression}{Signature matrix of cell type-specific signals. 
If not provided, can be computed from a provided ExpressionSet containing 
single-cell data.}

\item{cellScaleFactors}{Cell size scale factor transformations of 
length equal to the K cell types to deconvolve.}

\item{returnInfo}{Whether to return metadata and original method outputs 
with predicted proportions.}
}
\value{
New object.
}
\description{
Function to make a new object of class \linkS4class{independentbulkParam}
}
\examples{
new("independentbulkParam")

}
