% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_metadata.R
\name{luteSupportedDeconvolutionAlgorithms}
\alias{luteSupportedDeconvolutionAlgorithms}
\title{luteSupportedDeconvolutionAlgorithms}
\usage{
luteSupportedDeconvolutionAlgorithms()
}
\value{
Table of supported deconvolution algorithms.
}
\description{
View details about supported deconvolution algorithms.
}
\examples{
luteSupportedDeconvolutionAlgorithms()

}
