% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{make_new_assaydata}
\alias{make_new_assaydata}
\title{convert_sce}
\usage{
make_new_assaydata(ma, method = "mean", na.rm = TRUE, verbose = FALSE)
}
\arguments{
\item{ma}{Assay data matrix (cols = samples/cells/donors, rows = genes/loci).
Accepts tables of type data.frame, matrix, or DelayedArray.}

\item{method}{Statistical method to perform summary.}

\item{na.rm}{Option for NA removal.}

\item{verbose}{Whether to show verbose status messages.}

\item{make_new_assaydata}{}
}
\description{
Manage conversions to a `SummarizedExperimentTypes` object.
}
