% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_mappings.R
\name{map_deconrnaseq}
\alias{map_deconrnaseq}
\title{map_deconrnaseq}
\usage{
map_deconrnaseq(
  arguments,
  method = "DeconRNASeq",
  library.name = "DeconRNASeq",
  method.arguments = c(signatures = "as.data.frame(Z)", datasets =
    "as.data.frame(cbind(Y,Y))", use.scale = "FALSE")
)
}
\arguments{
\item{arguments}{List of user-provided arguments.}

\item{method}{Character string of the method.}

\item{library.name}{Name of library to call function from.}

\item{method.arguments}{Arguments required for this method.}
}
\value{
List of data and command character string to parse.
}
\description{
Mapping provided arguments to required arguments for DeconRNASeq, using 
defaults for any required arguments not provided
}
