% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music2Param-class.R
\docType{class}
\name{music2Param-class}
\alias{music2Param-class}
\alias{MuSiC2Param-class,}
\alias{Music2Param-class}
\title{music2Param-class}
\description{
Main constructor for class to manage mappings to the deconvolution
method function \code{MuSiC::music2()}.
}
\examples{
lexample <- .get_decon_example_data()

}
