% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicParam-class.R
\docType{class}
\name{musicParam-class}
\alias{musicParam-class}
\title{Run MuSiC}
\description{
Applies the MuSiC::music.basic() implementation of the MuSiC deconvolution 
algorithm.
}
\examples{
# example
lexample <- .get_decon_example_data()
param <- musicParam(s = lexample[["s"]], y = lexample[["y"]], z = lexample[["z"]])

# return only predicted proportions
deconvolution(param)
# type1     type2 
# 0.6770833 0.3229167

# return full results
param@return.info <- T
names(deconvolution(param))
# [1] "predictions" "result.info" "metadata"

}
