% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_rnf.R
\name{new_workflow_table}
\alias{new_workflow_table}
\title{new_workflow_table}
\usage{
new_workflow_table(
  singleCellExperimentNames = NULL,
  dataDirectory = "data",
  trueProportionsFilenameStem = "true_proportions_",
  cellTypeVariable = "celltype",
  tableDirectory = ".",
  tableFileName = "workflow_table.csv",
  save = TRUE,
  overwrite = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{singleCellExperimentNames}{Names of SingleCellExperiment files to load.}

\item{dataDirectory}{Directory containing datasets to load.}

\item{trueProportionsFilenameStem}{File name stem of true proportions values.}

\item{cellTypeVariable}{Name of variable containing cell type labels.}

\item{tableDirectory}{Directory to write table.}

\item{tableFileName}{The file name of the new table to write.}

\item{save}{Whether to save the new table.}

\item{overwrite}{Whether to overwrite old table files.}

\item{verbose}{Whether to show verbose messages (T/F).}
}
\value{
New r-nf_deconvolution compatible table of experiment/run metadata.
}
\description{
Makes a new experiment table for r-nf_deconvolution runs.
}
\details{
Makes and returns/saves a r-nf_deconvolution experiment table. 
Checks for existence of provided files.
}
\examples{
new_workflow_table(save=FALSE)

}
