% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{pca_bydonortype}
\alias{pca_bydonortype}
\title{pca_bydonortype}
\usage{
pca_bydonortype(
  dt,
  test.md = list(test = "pca", test.type = "by donor;type"),
  title.append = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dt}{Data table containing the donors (rows) by marker signals (columns)
for each marker and type.}

\item{test.md}{Metadata info to be returned with test results}

\item{title.append}{Optional string to append to plot titles.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
lr, results list containing PCA results data, ggplots, and metadata.
}
\description{
Get analysis results and plots for PCA by donor and type, across markers.
}
