% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcheffects_simulation-utilities.R
\name{pcaplots_donor}
\alias{pcaplots_donor}
\title{pcaplots_donor}
\usage{
pcaplots_donor(donordf, title.append = NULL, verbose = FALSE, ...)
}
\arguments{
\item{donordf}{Table of class `donor.data.frame`.}

\item{title.append}{Optional string to append to plot titles.}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Additional arguments passed to PCA functions.}
}
\value{
list of PCA results, plots, and metadata
}
\description{
Make plots of two PCAs: (1) by donor, across markers and types; (2) by donor
and type, across markers.
}
