% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{pdiff}
\alias{pdiff}
\title{pdiff}
\usage{
pdiff(pi, P, verbose = FALSE)
}
\arguments{
\item{pi}{Prediction set vector. Indices correspond to those in `P`.}

\item{P}{Prediction set vector. Indices correspond to those in `pi`.}

\item{verbose}{Whether to show verbose status updates.}
}
\value{
Comparisons across types, including bias (prediction - true), the
root mean squared error (RMSE), and Pearson and Spearman correlation 
coefficients.
}
\description{
Compare two sets of type predictions across types within an iteration.
}
\examples{
# examples
}
