% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-matrix-methods.R
\name{plot_dispersion_coef}
\alias{plot_dispersion_coef}
\title{plot_dispersion_coef}
\usage{
plot_dispersion_coef(
  dfp,
  make.boxplot = TRUE,
  make.jitter = TRUE,
  box.zoom.ymax = c(350, 50),
  jitter.zoom.ymax = c(350, 50),
  verbose = FALSE
)
}
\arguments{
\item{dfp}{Data.frame for plotting.}

\item{make.boxplot}{Whether to make ggplot boxplots.}

\item{make.jitter}{Whether to make ggplot jitter plots.}

\item{box.zoom.ymax}{Vector of y-axis maxima for two zoomed panels.}

\item{jitter.zoom.ymax}{Vector of y-axiz maxima for two zoomed panels.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
List of ggplot plot objects.
}
\description{
Plot dispersion coefficients by gene groups and cell types.
}
\details{
Makes summary plots of dispersion point estimates. Boxplots and 
jitter plots are generated at 3 zoom levels using the `ggforce::facet_zoom()` 
function. These are: 1. not zoomed; 2. zoom level 1; and 3. zoom level 2. The 
first and second zoom levels are the maximal y-axis values as specified by 
the arguments `box.zoom.ymax` and `jitter.zoom.ymax` for boxplots and jitter 
plots, respectively.
}
\seealso{
analyze_dispersion_est
}
