% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerdf_methods.R
\name{plot_marker_blockplot}
\alias{plot_marker_blockplot}
\title{plot_marker_blockplot}
\usage{
plot_marker_blockplot()
}
\description{
Plot a block plot of marker set overlaps using ComplexHeatmap annotations
objects.
}
