% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{rand_donor_marker_table}
\alias{rand_donor_marker_table}
\title{rand_donor_marker_table}
\usage{
rand_donor_marker_table(
  ndonor = 2,
  gindexv = c(1, 2),
  lambda.pos = 20,
  lambda.neg = 2,
  sd.offset.pos = 10,
  sd.offset.neg = 2,
  seed.num = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{ndonor}{Number of donors to simulate.}

\item{gindexv}{Vector of marker indices. Index values correspond to the k types,
and each index position represents a marker (e.g. c(1,2,2) means two markers 
for the second type, etc.).}

\item{lambda.pos}{Value of lambda (Poisson dist. mean) for "positive" marker 
status (e.g. mean of dist. for k when marker is positive for k, negative for 
not-k).}

\item{lambda.neg}{Value of lambda (Poisson dist. mean) for "negative" marker 
status (e.g. mean of dist. for k when marker is positive for not-k, negative 
for k).}

\item{sd.offset.pos}{Poisson dist mean for randomization of offsets for
positive marker signals.}

\item{sd.offset.neg}{Poisson dist mean for randomization of offsets for
negative marker signals.}

\item{seed.num}{Token to set the random seed.}

\item{...}{Additional parameters passed to `random_lgv()` to get marker 
signals.}

\item{vebose}{Whether to return verbose status messages.}
}
\value{
return
}
\description{
Get a flat table of random donor marker signals by types.
}
\examples{

rand_donor_marker_table(ndonor = 2, gindexv = c(1,2))

rand_donor_marker_table(ndonor = 10, gindexv = c(1,1,2))

rand_donor_marker_table(ndonor = 10, gindexv = c(rep(1, 10), rep(2, 20)))

}
\seealso{
random_lgv
}
