% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomized_data.R
\name{random_lgv}
\alias{random_lgv}
\title{random_lgv}
\usage{
random_lgv(
  gindexv,
  num.iter = 1,
  lambda.pos = 25,
  lambda.neg = 2,
  method = "nbinom",
  gamma.size.pos = 10,
  gamma.size.neg = 10,
  seed.num = 0
)
}
\arguments{
\item{gindexv}{Vector of marker indices. Index values correspond to the k types,
and each index position represents a marker (e.g. c(1,2,2) means two markers 
for the second type, etc.).}

\item{num.iter}{Total simulation iterations.}

\item{lambda.pos}{Value of lambda (Poisson dist. mean) for "positive" marker 
status (e.g. mean of dist. for k when marker is positive for k, negative for 
not-k). This is passed to the argument mu when method is "nbinom".}

\item{lambda.neg}{Value of lambda (Poisson dist. mean) for "negative" marker 
status (e.g. mean of dist. for k when marker is positive for not-k, negative 
for k). This is passed to the argument mu when method is "nbinom".}

\item{method}{Type of randomization method to use. Accepts either "poisson"
for poisson distribution (see `?rpois` for details), or "nbinom" for the 
negative binomial (a.k.a. gamm poisson) distribution (see `?rnbinom` for 
details).}

\item{gamma.size.pos}{The gamma distribution magnitude for "positive" markers. 
This is applied when the "nbinom" method is used.}

\item{gamma.size.neg}{The gamma distribution magnitude for "negative" markers. 
This is applied when the "nbinom" method is used.}

\item{seed.num}{Seed value for randomization.}
}
\value{
Listed lgv object containing the randomized marker values across 
types.
}
\description{
Get randomized markers using Poisson distribution sampling. For a given K,
we assume "positive" markers have higher values than for non-K types, and 
thus we sample from 2 different Poisson distributions defined by different
lambda values (e.g. arguments lambda.pos, lambda.neg). WE also use argument 
gindexv to define total markers as length(gindexv) and the marker balance as
relative counts of each type index.
}
\details{
For example, if gindex is c(1, 1, 2), we define 3 total markers, 2 positive
markers for type 1 (negative for type 2) and a single positive marker for 
type 2 (negative for type 1).
}
\examples{
set.seed(0)
random_lgv(gindexv = c(rep(1, 10), rep(2, 5)))
}
