% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomized_data.R
\name{random_sce}
\alias{random_sce}
\title{random_sce}
\usage{
random_sce(
  num.genes = 20,
  num.cells = 10,
  num.types = 2,
  expr.mean = 10,
  verbose = FALSE,
  seed.num = 0
)
}
\arguments{
\item{num.genes}{Number of genes to randomize.}

\item{num.cells}{Numnber of cells to randomize.}

\item{num.types}{Number of cell types to annotate.}

\item{expr.mean}{Poisson dist mean for random expression data.}

\item{verbose}{Whether to show verbose status messages.}

\item{seed.num}{Seed value for randomization of expression data.}
}
\value{
New randomized SingleCellExperiment object.
}
\description{
Make a random SingleCellExperiment object.
}
