% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencebasedParam-class.R
\docType{class}
\name{referencebasedParam-class}
\alias{referencebasedParam-class}
\title{referencebasedParam-class}
\value{
New object.
}
\description{
Class and methods for managing reference-based deconvolution methods.
}
\details{
This is a parent class to manage reference-based deconvolution 
algorithms. 

Child/sub-classes of this are distinguished by their use of
either an explicit or implied \code{z} signature matrix (i.e. Z[G,K] of
dimensions G markers by K cell types). These also have an implied cell size 
term for biases from systematic cell size differences. If no cell size 
transformation is intended, this is the equivalent of passing equal size 
scales, (e.g. a K-length vector of equal values). See 
`vignette(package="lute")` for details about experiment terms.
}
\examples{
lexample <- get_decon_example_data()
referencebasedParam(y = lexample$y, z = lexample$z, s = lexample$s)

}
