% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencebasedParam-class.R
\name{referencebasedParam}
\alias{referencebasedParam}
\title{Make new object of class referencebasedParam}
\usage{
referencebasedParam(
  bulkExpression,
  referenceExpression,
  cellScaleFactors,
  returnInfo = FALSE
)
}
\arguments{
\item{bulkExpression}{Bulk mixed signals matrix of samples, which can be 
matched to single-cell samples.}

\item{referenceExpression}{Signature matrix of cell type-specific signals. 
If not provided, can be computed from a provided ExpressionSet containing 
single-cell data.}

\item{cellScaleFactors}{Cell size factor transformations of length equal to 
the K cell types to deconvolve.}

\item{returnInfo}{Whether to return metadata and original method outputs 
with predicted proportions.}
}
\value{
New object of class \linkS4class{referencebasedParam}.

New object.
}
\description{
Main constructor for class \linkS4class{referencebasedParam}.
}
\details{
Takes standard inputs for reference-based deconvolution algorithms.
}
\examples{
exampleList <- getDeconvolutionExampleData()
referencebasedParam(
  bulkExpression=exampleList$bulkExpression, 
  referenceExpression=exampleList$referenceExpression, 
  cellScaleFactors=exampleList$cellScaleFactors
)

}
