% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencebasedParam-class.R
\name{referencebasedParam}
\alias{referencebasedParam}
\title{Function to get nnlsParam}
\usage{
referencebasedParam(y, z, s = NULL, return.info = FALSE)
}
\description{
Function to get nnlsParam
}
