% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencebasedParam-class.R
\name{referencebasedParam}
\alias{referencebasedParam}
\title{Make new object of class referencebasedParam}
\usage{
referencebasedParam(y, z, s, return.info = FALSE)
}
\arguments{
\item{y}{Bulk mixed signals matrix of samples, which can be matched to single-cell samples.}

\item{z}{Signature matrix of cell type-specific signals. If not provided, can be computed from a
provided ExpressionSet containing single-cell data.}

\item{s}{Cell size factor transformations of length equal to the K cell types to deconvolve.}

\item{return.info}{Whether to return metadata and original method outputs with predicted proportions.}
}
\value{
New object of class \linkS4class{referencebasedParam}.

New object.
}
\description{
Main constructor for class \linkS4class{referencebasedParam}.
}
\details{
Takes standard inputs for reference-based deconvolution algorithms.
}
\examples{
lexample <- get_decon_example_data()
referencebasedParam(y = lexample$y, z = lexample$z, s = lexample$s)

}
