% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencefreeParam-class.R
\docType{class}
\name{referencefreeParam-class}
\alias{referencefreeParam-class}
\title{referencefreeParam-class}
\description{
Class and methods for managing reference-free deconvolution methods.
}
\examples{
lexample <- .get_decon_example_data()

}
