% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{results_plots}
\alias{results_plots}
\title{results_plots}
\usage{
results_plots(dfres, lsv = NULL, refline.color = "red", verbose = FALSE)
}
\arguments{
\item{dfres}{Data.frame of deconvolution simulation results.}

\item{lsv}{List of size factor values. If length(lsv) > length(lpv), only use
up to the number of iterations in lpv. If NULL, S-transformation experiments
aren't performed, and Y is calculated as $P*Z$ rather than $P*ZS$.}

\item{refline.color}{Color of the reference line for the second scatterplot 
of RMSE by experiment type.}

\item{verbose}{Whether to show verbose status updates.}
}
\value{
List of ggplot2 objects analyzing deconvolution simulation results, 
including scatter plots and a violin plot.
}
\description{
Makes standard plots to analyze deconvolution simulation results.
}
\examples{
# example
}
\seealso{
decon_analysis
}
