% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_rmse.R
\name{rmseTest}
\alias{rmseTest}
\title{rmseTest}
\usage{
rmseTest(firstVector, secondVector)
}
\arguments{
\item{vector1}{First numeric vector}

\item{vector2}{Second numeric vector}
}
\value{
single numeric
}
\description{
Takes 2 vectors of numerics
}
\details{
Function to test RMSE values (see `test_rmse.R`)
}
\examples{
proportionsVectorPred <- seq(1e-10,2e-10,1e-11)
proportionsVectorTrue <- rev(proportionsVectorPred)
rmseTest(proportionsVectorTrue, proportionsVectorPred)

}
