% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_mappings.R
\name{run_deconvolution}
\alias{run_deconvolution}
\title{run_deconvolution}
\usage{
run_deconvolution(
  Z = NULL,
  Y = NULL,
  seed.num = 0,
  method = "nnls",
  arguments = list()
)
}
\arguments{
\item{Z}{Signature matrix of dimensions G (marker genes) x K (types).}

\item{Y}{Bulk matrix of dimensions G (marker genes) x J (bulk samples).}

\item{seed.num}{Random seed for computational reproducibility.}

\item{method}{Character string of a valid deconvolution method to use (see
available methods with \code{valid_deconvolution_methods()}).}

\item{arguments}{List of additional valid arguments for the method.}
}
\value{
List of objects of type \code{deconvolution.results} containing 
predictions, metadata, and benchmarking data, organized by J samples in the
provided Y bulk data.
}
\description{
Perform deconvolution to predict cell type proportions in mixed samples by
passing standard inputs and getting standard outputs across a number of 
different deconvolution methods/functions.
}
\details{
This is the main function to estimate cell type proportions using 
numerous different reference-based deconvolution methods. It calls several
other functions and provides a standard output from supported methods. 
Currently, one bulk/mixed signals sample is processed at a time. If multiple
bulk samples are provided (e.g. \code{ncol(Y) > 1}), then results for each
sample are returned in a named list of \code{deconvolution.results} objects.
}
\examples{
sce <- random_sce()
typev <- unique(sce[["celltype"]])
Z <- do.call(cbind, lapply(typev, function(typei){
rowMeans(counts(sce[,sce[["celltype"]]==typei]))
}))
colnames(Z) <- c('type1', 'type2')
Y <- matrix(rowMeans(counts(sce)), ncol = 1)

# run methods
# run nnls
ldecon <- run_deconvolution(method = "nnls", Y = Y, Z = Z)

# inspect results
ldecon

}
\seealso{
\code{\link{map_deconvolution_arguments}}, for the mapping function

\code{\link{get_deconvolution_predictions}}, for the prediction management 
function

\code{\link{map_nnls}}, for the default method maps
}
