% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdcParam-class.R
\docType{class}
\name{scdcParam-class}
\alias{scdcParam-class}
\alias{SCDCParam-class,}
\alias{ScdcParam-class}
\title{scdcParam-class}
\description{
Main constructor for class to manage mappings to the deconvolution
method function \code{SCDC::SCDC_prop()}.
}
\examples{
lexample <- .get_decon_example_data()

}
