% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{sce_groupstat}
\alias{sce_groupstat}
\title{sce_groupstat}
\usage{
sce_groupstat(
  scef,
  groupvar,
  ugroupv,
  assayname = "counts",
  summarytype = "rowData",
  groupstat = c("mean", "median", "var", "sd", "numzero"),
  verbose = FALSE
)
}
\arguments{
\item{scef}{Filtered `SingleCellExperiment` object. Should reflect the data
for a specific type.}

\item{groupvar}{Variable containing group information on scef colData.}

\item{ugroupv}{Vector of all unique groups to consider, of which levels in
groupvar entries should be identical or a subset.}

\item{summarytype}{Whether to summarize data on rowData (e.g. one entry per 
row/gene), or otherwise use the colData collapsed on means (e.g. take the 
mean expression across cells for each gene, returning one row per type).}

\item{groupstat}{Summary statistics to calculate. Can be either of "mean",
"median", "var", "sd", or "numzero" (i.e. number of zero-value entries).}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
data.frame containing group-level summary statistics for all groups 
specified in ugroupv.
}
\description{
Get group-level summary statistics, either on rowData or collapsed colData.
}
\examples{
sce <- random_sce()
colData(sce)$donor <- c(rep("donor1", 7), rep("donor2", 3))
dfg <- sce_groupstat(sce, "donor", c("group1", "group2"), "counts", "colData")
}
