% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_conversions.R
\name{se_to_eset}
\alias{se_to_eset}
\title{se_to_eset}
\usage{
se_to_eset(summarizedExperiment, assayName = "counts")
}
\arguments{
\item{summarizedExperiment}{Object of type SummarizedExperiment (see 
\code{?SummarizedExperiment}).}

\item{assayName}{Name of assay to store in new ExpressionSet object.}
}
\value{
New object of type ExpressionSet.
}
\description{
Convert SummarizedExperiment to ExpressionSet.
}
\examples{
summarizedExperiment <- sce_to_se(random_sce())
se_to_eset(summarizedExperiment)

}
