% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_conversions.R
\name{se_to_sce}
\alias{se_to_sce}
\title{se_to_sce}
\usage{
se_to_sce(summarizedExperiment)
}
\arguments{
\item{summarizedExperiment}{Object of type SummarizedExperiment (see 
\code{?SummarizedExperiment}).}
}
\value{
New SingleCellExperiment object.
}
\description{
Convert SummarizedExperiment to SingleCellExperiment.
}
\examples{
se_to_sce(SummarizedExperiment())

}
