% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_conversions.R
\name{se_to_sce}
\alias{se_to_sce}
\title{se_to_sce
Convert SummarizedExperiment to SingleCellExperiment.}
\usage{
se_to_sce(se)
}
\arguments{
\item{se}{SummarizedExperiment.}
}
\value{
SingleCellExperiment.
}
\description{
se_to_sce
Convert SummarizedExperiment to SingleCellExperiment.
}
\examples{
se_to_sce(SummarizedExperiment())

}
