% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperimentTypes-class.R
\name{set_from_sce}
\alias{set_from_sce}
\title{set_from_sce}
\usage{
set_from_sce(
  sce,
  groupvar = NULL,
  method = "mean",
  typevar = "celltype",
  assayname = "counts",
  verbose,
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{groupvar}{Variable containing group-level labels using in group 
summaries. If NULL, skip group summaries.}

\item{method}{Method to summarize assays on types.}

\item{typevar}{Variable containing the type-level labels.}

\item{assayname}{Name of assays data in assays(sce).}

\item{verbose}{Whether to show verbose status messages.}

\item{...}{Additional arguments specified for `sce_groupstat()` summaries by 
groups (see `?sce_groupstat` for details).}
}
\value{
SummarizedExperimentTypes object.
}
\description{
Make SummarizedExperimentTypes from SingleCellExperiment.
}
