% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{set_from_sce}
\alias{set_from_sce}
\title{set_from_sce}
\usage{
set_from_sce(
  sce,
  method = "mean",
  type.variable = "celltype",
  assayname = "counts",
  make.set.plots = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sce}{A `SingleCellExperiment` object.}

\item{method}{Method to summarize assays on types.}

\item{type.variable}{Variable containing the type-level labels.}

\item{assayname}{Name of assays data in assays(sce).}

\item{make.set.plots}{Make standard plots of SummarizedExperimentTypes, 
including a heatmap of type-level summaries.}

\item{verbose}{Whether to show verbose status messages.
groups (see `?sce_groupstat` for details).}
}
\value{
`SummarizedExperimentTypes` object.
}
\description{
Make an object of class `SummarizedExperimentTypes` from an object of class
`SingleCellExperiment` or `SummarizedExperiment`.
}
\details{
Get a SummarizedExperimentTypes object as a set of summary 
statistics applied on the indicated types variable across cells from a 
SingleCellExperiment object.

Type-level summaries are obtained rapidly using the function 
`scuttle::aggregateAcrossCells()`. Additional summary statistics, such as 
count summaries of zero-value data, are then calculated using calls to 
`sce_groupstat()`.
}
\examples{
sce <- random_sce()
set <- set_from_sce(sce)
show(set)
}
\seealso{
sce_groupstat
}
