% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{set_from_set}
\alias{set_from_set}
\title{set_from_set}
\usage{
set_from_set(
  set,
  group.variable = "donor",
  type.variable = "celltype",
  method = "mean",
  assayname = "counts_bytype",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{set}{An object of type SummarizedExperimentTypes or similar.}

\item{group.variable}{Name of variable containing group level information.}

\item{type.variable}{Name of variable containing type level information.}

\item{method}{Statistical method to summarize types on group levels for new
assays. Can be either of "mean" or "median".}

\item{assayname}{Name of assays data in provided set object to summarize.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
New set object with type data summarized on provided groups.
}
\description{
Expects a set object with type;group nesting. Returns a set object summarized
by each type across any available groups.
}
\examples{
sce <- random_sce()
sce[["donor"]] <- c(rep("donor1", 2), rep("donor2", 8))
sce[["type.variable"]] <- paste0(sce[["celltype"]], ";", sce[["donor"]])
set <- set_from_sce(sce, group.variable = "donor", type.variable = "celltype")
set2 <- set_from_set(set, type.variable = "celltype", group.variable = "donor")
}
