% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisqueParam-class.R
\name{show,bisqueParam-method}
\alias{show,bisqueParam-method}
\title{Show generic behavior for object of class bisqueParam}
\usage{
\S4method{show}{bisqueParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{bisqueParam} (see 
\code{?bisqueParam}).}
}
\value{
Prints data summary messages to console.
}
\description{
Show generic behavior for object of class bisqueParam
}
\examples{
## get data
exampleList <- getDeconvolutionExampleDataBisque()
bulkExpressionSet <- exampleList[["bulkExpressionSet"]][,seq(10)]
bulkExpression <- exprs(exampleList[["bulkExpressionSet"]])
bulkExpression <- bulkExpression[,c(11:ncol(bulkExpression))]

## get param object
newBisqueParameter <- bisqueParam(bulkExpressionSet=bulkExpressionSet, 
                     bulkExpressionIndependent=bulkExpression,
                     scData=exampleList[["singleCellExpressionSet"]], 
                     batchVariable="SubjectName", 
                     cellTypeVariable="cellType", 
                     useOverlap=FALSE)
## show
newBisqueParameter

}
