% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_classes.R
\name{show,cellProportionsPredictions-method}
\alias{show,cellProportionsPredictions-method}
\title{Inspect cellProportionsPredictions object.}
\usage{
\S4method{show}{cellProportionsPredictions}(object)
}
\arguments{
\item{object}{cellProportionsPredictions object.}
}
\value{
Shows object summaries.
}
\description{
Inspect cellProportionsPredictions object.
}
\details{
Method behavior for show.
}
\examples{
example.data <- get_decon_example_data()

}
