% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnlsParam-class.R
\name{show,nnlsParam-method}
\alias{show,nnlsParam-method}
\title{Show generic behavior for object of class nnlsParam}
\usage{
\S4method{show}{nnlsParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{nnlsParam} (see 
\code{?nnlsParam}).}
}
\value{
Prints data summary messages to console.
}
\description{
Show generic behavior for object of class nnlsParam
}
\examples{
exampleList <- getDeconvolutionExampleData()
referencebasedParam(
bulkExpression=exampleList$bulkExpression, 
referenceExpression=exampleList$referenceExpression, 
cellScaleFactors=exampleList$cellScaleFactors)

}
