% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnlsParam-class.R
\name{show,nnlsParam-method}
\alias{show,nnlsParam-method}
\title{Show generic behavior for object of class \linkS4class{nnlsParam}}
\usage{
\S4method{show}{nnlsParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{nnlsParam} (see 
\code{?nnlsParam}).}
}
\value{
Shows object summaries.
}
\description{
Show generic behavior for object of class \linkS4class{nnlsParam}
}
\details{
Method for behavior of show generic when called for object of class 
\linkS4class{nnlsParam}
}
\examples{
lexample <- get_decon_example_data()
param <- nnlsParam(s=lexample[["s"]], y=lexample[["y"]], 
                    z=lexample[["z"]])

## return only predicted proportions
deconvolution(param)

# return full results
param@return.info <- TRUE
names(deconvolution(param))

}
