% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencebasedParam-class.R,
%   R/nnlsParam-class.R
\name{show,referencebasedParam-method}
\alias{show,referencebasedParam-method}
\title{Show generic behavior for object of class referencebasedParam}
\usage{
\S4method{show}{referencebasedParam}(object)

\S4method{show}{referencebasedParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{referencebasedParam} (see 
\code{?referencebasedParam}).}
}
\value{
Prints data summary messages to console.

Prints data summary messages to console.
}
\description{
Show generic behavior for object of class referencebasedParam

Show generic behavior for object of class referencebasedParam
}
\examples{
exampleList <- getDeconvolutionExampleData()
referencebasedParam(
bulkExpression=exampleList$bulkExpression, 
referenceExpression=exampleList$referenceExpression, 
cellScaleFactors=exampleList$cellScaleFactors)

exampleList <- getDeconvolutionExampleData()
referencebasedParam(
bulkExpression=exampleList$bulkExpression, 
referenceExpression=exampleList$referenceExpression, 
cellScaleFactors=exampleList$cellScaleFactors)

}
