% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencebasedParam-class.R
\name{show,referencebasedParam-method}
\alias{show,referencebasedParam-method}
\title{Show generic behavior for object of class referencebasedParam}
\usage{
\S4method{show}{referencebasedParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{referencebasedParam} (see 
\code{?referencebasedParam}).}
}
\value{
Prints data summary messages to console.
}
\description{
Show generic behavior for object of class referencebasedParam
}
\examples{
exampleList <- get_decon_example_data()
referencebasedParam(
bulkExpression=exampleList$bulkExpression, 
referenceExpression=exampleList$referenceExpression, 
cellScaleFactors=exampleList$cellScaleFactors)

}
