% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_donor_utils.R
\name{simulate_sce_donor_bias}
\alias{simulate_sce_donor_bias}
\title{simulate_sce_donor_bias}
\usage{
simulate_sce_donor_bias(donor.coeff.vector = NULL)
}
\arguments{
\item{donor.coeff.vector}{Vector of donor bias values. If NULL, uses 
`rnorm(10, mean = 200, sd = 30)`.}
}
\value{
List containing the sce and ggplot objects.
}
\description{
Simulate and plot donor bias from a SingleCellExperiment.
}
\details{
Simulates a new SingleCellExperiment object with included donor bias,
and plots the means and variances on log scale to show dispersion biases 
across simulated donors.
}
