% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor-subsample_utilities.R
\name{subsample_summary}
\alias{subsample_summary}
\title{subsample_summary}
\usage{
subsample_summary(results.table)
}
\arguments{
\item{results.table}{Table of subsample experiment results.}
}
\value{
dfs, data.frame containing summary statistics.
}
\description{
Get summary statistics for a results table from a subsampling experiment.
}
\examples{
path <- file.path("inst","examples","results-table.csv")
path <- system.file("./inst/examples/results-table.csv", package = 'lute')
results.table <- read.csv(path)
dfstat <- subsample_summary(results.table)
}
