% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmarkersParam-class.R
\name{typemarkers,findmarkersParam-method}
\alias{typemarkers,findmarkersParam-method}
\title{Cell type markers method for findmarkersParam}
\usage{
\S4method{typemarkers}{findmarkersParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{findmarkersParam} (see 
\code{?findmarkersParam}).}
}
\value{
Returns the top available markers, with type-specific marker filters,
as either a vector of marker IDs or a results list.
}
\description{
Defines the typemarkers method for \linkS4class{findmarkersParam}.
}
\details{
Takes an object of class \linkS4class{findmarkersParam} as input, 
returning either a vector of cell type gene markers, or (if 
\code{return.info == TRUE}) a list containing such a vector along with 
original function outputs.
}
\examples{
lexample <- get_decon_example_data()
sce.example <- random_sce()
new.param <- findmarkersParam(sce=sce.example, 
celltype.variable="celltype", markers.per.type=5)
markers <- typemarkers(new.param)

}
