% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typemarkersParam-class.R
\name{typemarkers,typemarkersParam-method}
\alias{typemarkers,typemarkersParam-method}
\title{Deconvolution method for class \linkS4class{deconrnaseqParam}}
\usage{
\S4method{typemarkers}{typemarkersParam}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{deconrnaseqParam}.}
}
\value{
Either a vector of gene markers, or a list of detailed outputs that
includes such a marker vector.
}
\description{
Main deconvolution method for the \linkS4class{deconrnaseqParam} to run the 
\code{DeconRNASeq::DeconRNASeq()} implementation of the DeconRNASeq algorithm.
}
