% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{ypb_from_sce}
\alias{ypb_from_sce}
\title{ypb_from_sce}
\usage{
ypb_from_sce(
  singleCellExperiment,
  assayName = "counts",
  cellTypeVariable = "celltype",
  sampleIdVariable = NULL,
  cellScaleFactors = NULL
)
}
\arguments{
\item{singleCellExperiment}{An object of type 
\linkS4class{SingleCellExperiment}.}

\item{assayName}{Name of expression matrix in \code{singleCellExperiment} 
assays.}

\item{cellTypeVariable}{Variable name for cell type labels in 
\code{singleCellExperiment} 
coldata.}

\item{sampleIdVariable}{Variable name for sample/group ID labels in 
\code{singleCellExperiment} coldata.}

\item{cellScaleFactors}{Vector of cell type size scale factors. Optional.}
}
\value{
Matrix of simulated bulk convoluted signals.
}
\description{
Get pseudobulk from a SingleCellExperiment object.
}
\examples{
singleCellExperimentExample <- random_singleCellExperiment()
ypb_from_sce(singleCellExperimentExample)

}
