% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{ypb_from_sce}
\alias{ypb_from_sce}
\title{ypb_from_sce}
\usage{
ypb_from_sce(
  sce,
  assay.name = "counts",
  celltype.variable = "celltype",
  sample.id.variable = NULL,
  S = NULL
)
}
\arguments{
\item{sce}{An object of type \linkS4class{SingleCellExperiment}.}

\item{assay.name}{Name of expression matrix in \code{sce} assays.}

\item{celltype.variable}{Variable name for cell type labels in \code{sce} 
coldata.}

\item{sample.id.variable}{Variable name for sample/group ID labels in 
\code{sce} coldata.}

\item{S}{Vector of cell type size scale factors. Optional.}
}
\value{
Matrix of simulated bulk convoluted signals.
}
\description{
Get pseudobulk from a SingleCellExperiment object.
}
\examples{
sce.example <- random_sce()
ypb_from_sce(sce.example)

}
