% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lute_utilities.R
\name{ypb_from_sce}
\alias{ypb_from_sce}
\title{ypb_from_sce}
\usage{
ypb_from_sce(sce, assay.name, celltype.variable, S = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{assay.name}{Valid assay name.}

\item{celltype.variable}{Valid cell type label in coldata.}

\item{S}{Vector of cell type size scale factors. Optional.}
}
\value{
Matrix of simulated bulk convoluted signals.
}
\description{
Get pseudobulk from a SingleCellExperiment object.
}
